Removing GVRContext content loading APis
This patch deprecates all the functions on GVRContext which load assets
(models, meshes, textures). It also moves the functions which load
meshes from GVRContext to GVRAssetLoader.
 Updated 3D cursor lib to not used deprecated asset APIs
