+++ GVRf/Framework/framework/src/main/java/org/gearvrf/GVRContext.java
@@ -60,13 +60,34 @@ import java.util.concurrent.Future;
 /**
  * Like the Android {@link Context} class, {@code GVRContext} provides core
  * services, and global information about an application environment.
- * 
- * Use {@code GVRContext} to {@linkplain #createQuad(float, float) create} and
- * {@linkplain #loadMesh(GVRAndroidResource) load} GL meshes, Android
- * {@linkplain #loadBitmap(String) bitmaps}, and
- * {@linkplain #loadTexture(GVRAndroidResource) GL textures.} {@code GVRContext}
- * also holds the {@linkplain GVRScene main scene} and miscellaneous information
- * like {@linkplain #getFrameTime() the frame time.}
+ * <ul>
+ * <li>
+ * The application <b>activity</b> resides in the context {@link #getActivity()}.
+ * </li>
+ * <li>
+ * The current scene in the context contains all the displayable 3D objects {@link #getMainScene()}.
+ * </li>
+ * <li>
+ * The <b>event receiver</b> in the context listens for events from scripting,
+ * picking, the asset loader and dispatches to user callbacks {@link #getEventReceiver()}.
+ * </li>
+ * <li>
+ * The context allows you to run code on either the Java or rendering thread
+ * {@link #runOnGlThread(Runnable), {@link #runOnTheFrameworkThread(Runnable)}.
+ * </li>
+ * <li>
+ * The <b>asset loader</b> in the context can load textures and models from a variety of sources
+ * both synchronously and asynchronously {@link #getAssetLoader()}.
+ * </li>
+ * <li>
+ * You can capture the 3D screen using context screen capture functions {@link #captureScreenCenter(GVRScreenshotCallback)}.
+ * </li>
+ * <li>
+ *  The <b>shader manager</b> in the context lets you create custom shaders.
+ * </ul>
+ * @see GVRAssetLoader
+ * @see GVREventReceiver
+ * @see GVRScene
  */
 public abstract class GVRContext implements IEventReceiver {
     private static final String TAG = Log.tag(GVRContext.class);
@@ -127,7 +148,7 @@ public abstract class GVRContext implements IEventReceiver {
 
     /**
      * The priority used by
-     * {@link #loadBitmapTexture(GVRAndroidResource.BitmapTextureCallback, GVRAndroidResource)}
+     * {@link GVRAssetLoader#loadTexture(GVRAndroidResource)}
      * and
      * {@link #loadMesh(GVRAndroidResource.MeshCallback, GVRAndroidResource)}
      *
@@ -242,588 +263,6 @@ public abstract class GVRContext implements IEventReceiver {
     public GVREventReceiver getEventReceiver() {
         return mEventReceiver;
     }
-    
-    /**
-     * Loads a file as a {@link GVRMesh}.
-     * 
-     * Note that this method can be quite slow; we recommend never calling it
-     * from the GL thread. The asynchronous version
-     * {@link #loadMesh(GVRAndroidResource.MeshCallback, GVRAndroidResource)} is
-     * better because it moves most of the work to a background thread, doing as
-     * little as possible on the GL thread.
-     * 
-     * @param androidResource
-     *            Basically, a stream containing a 3D model. The
-     *            {@link GVRAndroidResource} class has six constructors to
-     *            handle a wide variety of Android resource types. Taking a
-     *            {@code GVRAndroidResource} here eliminates six overloads.
-     * @return The file as a GL mesh or null if mesh cannot be loaded.
-     * 
-     * @since 1.6.2
-     */
-    public GVRMesh loadMesh(GVRAndroidResource androidResource) {
-        return loadMesh(androidResource,
-                GVRImportSettings.getRecommendedSettings());
-    }
-
-    /**
-     * Loads a file as a {@link GVRMesh}.
-     * 
-     * Note that this method can be quite slow; we recommend never calling it
-     * from the GL thread. The asynchronous version
-     * {@link #loadMesh(GVRAndroidResource.MeshCallback, GVRAndroidResource)} is
-     * better because it moves most of the work to a background thread, doing as
-     * little as possible on the GL thread.
-     * 
-     * @param androidResource
-     *            Basically, a stream containing a 3D model. The
-     *            {@link GVRAndroidResource} class has six constructors to
-     *            handle a wide variety of Android resource types. Taking a
-     *            {@code GVRAndroidResource} here eliminates six overloads.
-     * 
-     * @param settings
-     *            Additional import {@link GVRImportSettings settings}.
-     * @return The file as a GL mesh or null if mesh cannot be loaded.
-     * 
-     * @since 1.6.2
-     */
-    public GVRMesh loadMesh(GVRAndroidResource androidResource,
-            EnumSet<GVRImportSettings> settings)
-    {
-        class MeshFinder implements GVRSceneObject.ComponentVisitor
-        {
-            private GVRMesh meshFound = null;
-            public GVRMesh getMesh() { return meshFound; }
-            public boolean visit(GVRComponent comp)
-            {
-                GVRRenderData rdata = (GVRRenderData) comp;
-                meshFound = rdata.getMesh();
-                return (meshFound == null);
-            }
-        };
-        MeshFinder findMesh = new MeshFinder();
-        GVRMesh mesh = meshCache.get(androidResource);
-        if (mesh == null)
-        {
-            try
-            {
-                GVRSceneObject model = mImporter.loadModel(androidResource, settings, true, null);
-                model.forAllComponents(findMesh, GVRRenderData.getComponentType());
-                mesh = findMesh.getMesh();
-                if (mesh != null)
-                {
-                    meshCache.put(androidResource, mesh);
-                }
-                else
-                {
-                    throw new IOException("No mesh found in model " + androidResource.getResourceFilename());
-                }
-            }
-            catch (IOException ex)
-            {
-                getEventManager().sendEvent(this, IAssetEvents.class,
-                                            "onModelError", new Object[]{this, ex.getMessage(),
-                                androidResource.getResourceFilename()});
-                return null;
-            }
-        }
-        return mesh;
-    }
-
-    private final ResourceCache<GVRMesh> meshCache = new ResourceCache<GVRMesh>();
-
-    /**
-     * Loads a mesh file, asynchronously, at a default priority.
-     * 
-     * This method and the
-     * {@linkplain #loadMesh(GVRAndroidResource.MeshCallback, GVRAndroidResource, int)
-     * overload that takes a priority} are generally going to be your best
-     * choices for loading {@link GVRMesh} resources: mesh loading can take
-     * hundreds - and even thousands - of milliseconds, and so should not be
-     * done on the GL thread in either {@link GVRMain#onInit(GVRContext)}
-     * onInit()} or {@link GVRMain#onStep() onStep()}.
-     * 
-     * <p>
-     * The asynchronous methods improve throughput in three ways. First, by
-     * doing all the work on a background thread, then delivering the loaded
-     * mesh to the GL thread on a {@link #runOnGlThread(Runnable)
-     * runOnGlThread()} callback. Second, they use a throttler to avoid
-     * overloading the system and/or running out of memory. Third, they do
-     * 'request consolidation' - if you issue any requests for a particular file
-     * while there is still a pending request, the file will only be read once,
-     * and each callback will get the same {@link GVRMesh}.
-     * 
-     * @param callback
-     *            App supplied callback, with three different methods.
-     *            <ul>
-     *            <li>Before loading, GVRF may call
-     *            {@link GVRAndroidResource.MeshCallback#stillWanted(GVRAndroidResource)
-     *            stillWanted()} (on a background thread) to give you a chance
-     *            to abort a 'stale' load.
-     * 
-     *            <li>Successful loads will call
-     *            {@link GVRAndroidResource.Callback#loaded(GVRHybridObject, GVRAndroidResource)
-     *            loaded()} on the GL thread.
-     * 
-     *            <li>Any errors will call
-     *            {@link GVRAndroidResource.MeshCallback#failed(Throwable, GVRAndroidResource)
-     *            failed(),} with no promises about threading.
-     *            </ul>
-     * @param androidResource
-     *            Basically, a stream containing a 3D model. The
-     *            {@link GVRAndroidResource} class has six constructors to
-     *            handle a wide variety of Android resource types. Taking a
-     *            {@code GVRAndroidResource} here eliminates six overloads.
-     * @throws IllegalArgumentException
-     *             If either parameter is {@code null} or if you 'abuse' request
-     *             consolidation by passing the same {@link GVRAndroidResource}
-     *             descriptor to multiple load calls.
-     *             <p>
-     *             It's fairly common for multiple scene objects to use the same
-     *             texture or the same mesh. Thus, if you try to load, say,
-     *             {@code R.raw.whatever} while you already have a pending
-     *             request for {@code R.raw.whatever}, it will only be loaded
-     *             once; the same resource will be used to satisfy both (all)
-     *             requests. This "consolidation" uses
-     *             {@link GVRAndroidResource#equals(Object)}, <em>not</em>
-     *             {@code ==} (aka "reference equality"): The problem with using
-     *             the same resource descriptor is that if requests can't be
-     *             consolidated (because the later one(s) came in after the
-     *             earlier one(s) had already completed) the resource will be
-     *             reloaded ... but the original descriptor will have been
-     *             closed.
-     * @since 1.6.2
-     */
-    public void loadMesh(MeshCallback callback,
-            GVRAndroidResource androidResource) throws IllegalArgumentException {
-        loadMesh(callback, androidResource, DEFAULT_PRIORITY);
-    }
-
-    /**
-     * Loads a mesh file, asynchronously, at an explicit priority.
-     * 
-     * This method and the
-     * {@linkplain #loadMesh(GVRAndroidResource.MeshCallback, GVRAndroidResource)
-     * overload that supplies a default priority} are generally going to be your
-     * best choices for loading {@link GVRMesh} resources: mesh loading can take
-     * hundreds - and even thousands - of milliseconds, and so should not be
-     * done on the GL thread in either {@link GVRMain#onInit(GVRContext)
-     * onInit()} or {@link GVRMain#onStep() onStep()}.
-     * 
-     * <p>
-     * The asynchronous methods improve throughput in three ways. First, by
-     * doing all the work on a background thread, then delivering the loaded
-     * mesh to the GL thread on a {@link #runOnGlThread(Runnable)
-     * runOnGlThread()} callback. Second, they use a throttler to avoid
-     * overloading the system and/or running out of memory. Third, they do
-     * 'request consolidation' - if you issue any requests for a particular file
-     * while there is still a pending request, the file will only be read once,
-     * and each callback will get the same {@link GVRMesh}.
-     * 
-     * @param callback
-     *            App supplied callback, with three different methods.
-     *            <ul>
-     *            <li>Before loading, GVRF may call
-     *            {@link GVRAndroidResource.MeshCallback#stillWanted(GVRAndroidResource)
-     *            stillWanted()} (on a background thread) to give you a chance
-     *            to abort a 'stale' load.
-     * 
-     *            <li>Successful loads will call
-     *            {@link GVRAndroidResource.Callback#loaded(GVRHybridObject, GVRAndroidResource)
-     *            loaded()} on the GL thread.
-     * 
-     *            <li>Any errors will call
-     *            {@link GVRAndroidResource.MeshCallback#failed(Throwable, GVRAndroidResource)
-     *            failed(),} with no promises about threading.
-     *            </ul>
-     * @param resource
-     *            Basically, a stream containing a 3D model. The
-     *            {@link GVRAndroidResource} class has six constructors to
-     *            handle a wide variety of Android resource types. Taking a
-     *            {@code GVRAndroidResource} here eliminates six overloads.
-     * @param priority
-     *            This request's priority. Please see the notes on asynchronous
-     *            priorities in the <a href="package-summary.html#async">package
-     *            description</a>.
-     * 
-     * @throws IllegalArgumentException
-     *             If either {@code callback} or {@code resource} is
-     *             {@code null}, or if {@code priority} is out of range - or if
-     *             you 'abuse' request consolidation by passing the same
-     *             {@link GVRAndroidResource} descriptor to multiple load calls.
-     *             <p>
-     *             It's fairly common for multiple scene objects to use the same
-     *             texture or the same mesh. Thus, if you try to load, say,
-     *             {@code R.raw.whatever} while you already have a pending
-     *             request for {@code R.raw.whatever}, it will only be loaded
-     *             once; the same resource will be used to satisfy both (all)
-     *             requests. This "consolidation" uses
-     *             {@link GVRAndroidResource#equals(Object)}, <em>not</em>
-     *             {@code ==} (aka "reference equality"): The problem with using
-     *             the same resource descriptor is that if requests can't be
-     *             consolidated (because the later one(s) came in after the
-     *             earlier one(s) had already completed) the resource will be
-     *             reloaded ... but the original descriptor will have been
-     *             closed.
-     * @since 1.6.2
-     */
-    public void loadMesh(MeshCallback callback, GVRAndroidResource resource,
-            int priority) throws IllegalArgumentException {
-        GVRAsynchronousResourceLoader.loadMesh(this, callback, resource,
-                priority);
-    }
-
-    /**
-     * Simple, high-level method to load a mesh asynchronously, for use with
-     * {@link GVRRenderData#setMesh(Future)}.
-     * 
-     * This method uses a default priority; use
-     * {@link #loadFutureMesh(GVRAndroidResource, int)} to specify a priority;
-     * use one of the lower-level
-     * {@link #loadMesh(GVRAndroidResource.MeshCallback, GVRAndroidResource)}
-     * methods to get more control over loading.
-     * 
-     * @param resource
-     *            Basically, a stream containing a 3D model. The
-     *            {@link GVRAndroidResource} class has six constructors to
-     *            handle a wide variety of Android resource types. Taking a
-     *            {@code GVRAndroidResource} here eliminates six overloads.
-     * @return A {@link Future} that you can pass to
-     *         {@link GVRRenderData#setMesh(Future)}
-     * 
-     * @since 1.6.7
-     * 
-     * @throws IllegalArgumentException
-     *             If you 'abuse' request consolidation by passing the same
-     *             {@link GVRAndroidResource} descriptor to multiple load calls.
-     *             <p>
-     *             It's fairly common for multiple scene objects to use the same
-     *             texture or the same mesh. Thus, if you try to load, say,
-     *             {@code R.raw.whatever} while you already have a pending
-     *             request for {@code R.raw.whatever}, it will only be loaded
-     *             once; the same resource will be used to satisfy both (all)
-     *             requests. This "consolidation" uses
-     *             {@link GVRAndroidResource#equals(Object)}, <em>not</em>
-     *             {@code ==} (aka "reference equality"): The problem with using
-     *             the same resource descriptor is that if requests can't be
-     *             consolidated (because the later one(s) came in after the
-     *             earlier one(s) had already completed) the resource will be
-     *             reloaded ... but the original descriptor will have been
-     *             closed.
-     */
-    public Future<GVRMesh> loadFutureMesh(GVRAndroidResource resource) {
-        return loadFutureMesh(resource, DEFAULT_PRIORITY);
-    }
-
-    /**
-     * Simple, high-level method to load a mesh asynchronously, for use with
-     * {@link GVRRenderData#setMesh(Future)}.
-     * 
-     * This method trades control for convenience; use one of the lower-level
-     * {@link #loadMesh(GVRAndroidResource.MeshCallback, GVRAndroidResource)}
-     * methods if, say, you want to do something more than just
-     * {@link GVRRenderData#setMesh(GVRMesh)} when the mesh loads.
-     * 
-     * @param resource
-     *            Basically, a stream containing a 3D model. The
-     *            {@link GVRAndroidResource} class has six constructors to
-     *            handle a wide variety of Android resource types. Taking a
-     *            {@code GVRAndroidResource} here eliminates six overloads.
-     * @param priority
-     *            This request's priority. Please see the notes on asynchronous
-     *            priorities in the <a href="package-summary.html#async">package
-     *            description</a>.
-     * @return A {@link Future} that you can pass to
-     *         {@link GVRRenderData#setMesh(Future)}
-     * 
-     * @since 1.6.7
-     * 
-     * @throws IllegalArgumentException
-     *             If you 'abuse' request consolidation by passing the same
-     *             {@link GVRAndroidResource} descriptor to multiple load calls.
-     *             <p>
-     *             It's fairly common for multiple scene objects to use the same
-     *             texture or the same mesh. Thus, if you try to load, say,
-     *             {@code R.raw.whatever} while you already have a pending
-     *             request for {@code R.raw.whatever}, it will only be loaded
-     *             once; the same resource will be used to satisfy both (all)
-     *             requests. This "consolidation" uses
-     *             {@link GVRAndroidResource#equals(Object)}, <em>not</em>
-     *             {@code ==} (aka "reference equality"): The problem with using
-     *             the same resource descriptor is that if requests can't be
-     *             consolidated (because the later one(s) came in after the
-     *             earlier one(s) had already completed) the resource will be
-     *             reloaded ... but the original descriptor will have been
-     *             closed.
-     */
-    public Future<GVRMesh> loadFutureMesh(GVRAndroidResource resource,
-            int priority) {
-        return GVRAsynchronousResourceLoader.loadFutureMesh(this, resource,
-                priority);
-    }
-
-    /**
-     * Simple, high-level method to load a scene as {@link GVRSceneObject} from
-     * 3D model.
-     * 
-     * @param assetRelativeFilename
-     *            A filename, relative to the {@code assets} directory. The file
-     *            can be in a sub-directory of the {@code assets} directory:
-     *            {@code "foo/bar.png"} will open the file
-     *            {@code assets/foo/bar.png}
-     * 
-     * @return A {@link GVRSceneObject} that contains the meshes with textures
-     * 
-     * @throws IOException
-     *             File does not exist or cannot be read
-     *
-     * @deprecated Replaced by {@link #loadModel}
-     */
-    public GVRSceneObject getAssimpModel(String assetRelativeFilename)
-            throws IOException {
-        return mImporter.loadModel(assetRelativeFilename,
-                GVRImportSettings.getRecommendedSettings(), true, (GVRScene) null);
-    }
-
-    /**
-     * Simple, high-level method to load a scene as {@link GVRSceneObject} from
-     * 3D model.
-     * 
-     * @param assetRelativeFilename
-     *            A filename, relative to the {@code assets} directory. The file
-     *            can be in a sub-directory of the {@code assets} directory:
-     *            {@code "foo/bar.png"} will open the file
-     *            {@code assets/foo/bar.png}
-     * 
-     * @param settings
-     *            Additional import {@link GVRImportSettings settings}
-     * @return A {@link GVRSceneObject} that contains the meshes with textures
-     * 
-     * @throws IOException
-     *             File does not exist or cannot be read
-     *
-     * @deprecated Replaced by {@link GVRAssetLoader#loadModel(String)}
-     */
-    public GVRSceneObject getAssimpModel(String assetRelativeFilename,
-            EnumSet<GVRImportSettings> settings) throws IOException {
-        return mImporter.loadModel(assetRelativeFilename, settings, true, (GVRScene) null);
-    }
-    
-    /**
-     * @deprecated Replaced by {@link GVRAssetLoader#loadModel(String)}
-     */
-    public GVRModelSceneObject loadJassimpModelFromSD(String externalFile) throws IOException {
-        return mImporter.loadModel("sd:" + externalFile, GVRImportSettings.getRecommendedSettings(), true, null);
-    }
-
-    /**
-     * @deprecated Replaced by {@link GVRAssetLoader#loadModel(String)}
-     */
-    public GVRModelSceneObject loadJassimpModelFromSD(String externalFile, EnumSet<GVRImportSettings> settings) throws IOException {
-        return mImporter.loadModel("sd:" + externalFile, settings, true, null);
-    }
-
-    /**
-     * @deprecated Replaced by {@link GVRAssetLoader#loadModel(String)}
-     */
-    public GVRModelSceneObject loadJassimpModel(String assetFile) throws IOException {
-        return  mImporter.loadModel(assetFile, GVRImportSettings.getRecommendedSettings(), true, null);
-    }
-
-    /**
-     * @deprecated Replaced by {@link GVRAssetLoader#loadModel(String)}
-     */
-    public GVRModelSceneObject loadJassimpModel(String assetFile, EnumSet<GVRImportSettings> settings) throws IOException {
-        return mImporter.loadModel(assetFile, settings, true, null);
-    }
-
-    /**
-     * Simple, high-level method to load a scene object {@link GVRModelSceneObject} from
-     * a 3D model stored in the Android SD card.
-     *
-     * @param externalFile
-     *            A filename, relative to the SD card's root directory.
-     *
-     * @return A {@link GVRModelSceneObject} that contains the meshes with textures and bones
-     * and animations.
-     *
-     * @throws IOException
-     *             File does not exist or cannot be read
-     * @deprecated Replaced by {@link GVRAssetLoader#loadModel(String)}
-     */
-    public GVRModelSceneObject loadModelFromSD(String externalFile) throws IOException {
-        return mImporter.loadModel("sd:" + externalFile, GVRImportSettings.getRecommendedSettings(), true, null);
-    }
-
-    /**
-     * Simple, high-level method to load a scene object {@link GVRModelSceneObject} from
-     * a 3D model stored in the Android SD card.
-     *
-     * @param externalFile
-     *            A filename, relative to the SD card's root directory.
-     *
-     * @param settings
-     *            Additional import {@link GVRImportSettings settings}
-     *
-     * @return A {@link GVRModelSceneObject} that contains the meshes with textures and bones
-     * and animations.
-     *
-     * @throws IOException
-     *             File does not exist or cannot be read
-     * @deprecated Replaced by {@link GVRAssetLoader#loadModel(String)}
-     */
-    public GVRModelSceneObject loadModelFromSD(String externalFile, EnumSet<GVRImportSettings> settings) throws IOException {
-        return mImporter.loadModel("sd:" + externalFile, settings, true, null);
-    }
-
-    /**
-     * Simple, high-level method to load a scene object {@link GVRModelSceneObject} from a 3D model.
-     *
-     * @param assetFile The filename or URL of the model.
-     *                  If the filename begins with "sd:" it is assumed to be on the SD card.
-     *                  If it begins with "http:" or "https:", it is assumed to be a URL.
-     *                  Otherwise, it is assumed to be in the {@code assets directory}:
-     *                  {@code "foo/bar.png"} will open the file {@code assets/foo/bar.png}
-     *
-     * @return A {@link GVRModelSceneObject} that contains the meshes with textures and bones
-     * and animations.
-     *
-     * @throws IOException
-     *             File does not exist or cannot be read
-     * @deprecated use GVRAssetLoader.loadModel instead
-     */
-    public GVRModelSceneObject loadModel(String assetFile) throws IOException {
-        return mImporter.loadModel(assetFile, GVRImportSettings.getRecommendedSettings(), true, null);
-    }
-
-    /**
-     * Load a scene object {@link GVRModelSceneObject} from a 3D model and add it to the scene.
-     *
-     * @param assetFile The filename or URL of the model.
-     *                  If the filename begins with "sd:" it is assumed to be on the SD card.
-     *                  If it begins with "http:" or "https:", it is assumed to be a URL.
-     *                  Otherwise, it is assumed to be in the {@code assets directory}:
-     *                  {@code "foo/bar.png"} will open the file {@code assets/foo/bar.png}
-     *
-     * @param settings
-     *            Additional import {@link GVRImportSettings settings}
-     *            
-     * @param scene
-     *            If present, this asset loader will wait until all of the textures have been
-     *            loaded and then adds the model to the scene.
-     *            
-     * @return A {@link GVRModelSceneObject} that contains the meshes with textures and bones
-     * and animations.
-     *
-     * @throws IOException
-     *             File does not exist or cannot be read
-     * @deprecated use GVRAssetLoader.loadModel instead
-     *
-     */
-    public GVRModelSceneObject loadModel(String assetFile, EnumSet<GVRImportSettings> settings, GVRScene scene) throws IOException {
-        return mImporter.loadModel(assetFile, settings, true, scene);
-    }
-
-    /**
-     * Load a scene object {@link GVRModelSceneObject} from a 3D model.
-     *
-     * @param assetFile
-     *            A filename, relative to the {@code assets} directory. The file
-     *            can be in a sub-directory of the {@code assets} directory:
-     *            {@code "foo/bar.png"} will open the file
-     *            {@code assets/foo/bar.png}
-     *
-     * @param settings
-     *            Additional import {@link GVRImportSettings settings}
-     *            
-     * @return A {@link GVRModelSceneObject} that contains the meshes with textures and bones
-     * and animations.
-     *
-     * @throws IOException
-     *             File does not exist or cannot be read
-     * @deprecated use GVRAssetLoader.loadModel instead
-     */
-    public GVRModelSceneObject loadModel(String assetFile, EnumSet<GVRImportSettings> settings) throws IOException {
-        return mImporter.loadModel(assetFile, settings, true, null);
-    }
-
-    /**
-     * Simple, high-level method to load a scene object {@link GVRModelSceneObject} from
-     * a 3D model from a URL.
-     *
-     * @param urlString
-     *            A URL string pointing to where the model file is located.
-     *
-     * @return A {@link GVRModelSceneObject} that contains the meshes with textures and bones
-     * and animations.
-     *
-     * @throws IOException
-     *             File does not exist or cannot be read
-     * @deprecated use GVRAssetLoader.loadModel instead
-     *
-     */
-    public GVRSceneObject loadModelFromURL(String urlString) throws IOException {
-        return loadModelFromURL(urlString, false);
-    }
-
-    /**
-     * Simple, high-level method to load a scene object {@link GVRModelSceneObject} from
-     * a 3D model from a URL with an option to store it into local file cache for future use.
-     *
-     * @param urlString
-     *            A URL string pointing to where the model file is located.
-     *
-     * @param cacheEnabled
-     *           An option that a developer can choose for the loading process for the trade off 
-     *           between performance and consistency(or security):
-     *           True: Download and store the file locally in this app's cache directory and
-     *                  load from the cache in future uses;
-     *           false: Don't keep local copy and only do online streaming for loading everytime
-     *           
-     * @return A {@link GVRModelSceneObject} that contains the meshes with textures and bones
-     * and animations.
-     *
-     * @throws IOException
-     *             File does not exist or cannot be read
-     * @deprecated use GVRAssetLoader.loadModel instead
-     *
-     */
-    public GVRSceneObject loadModelFromURL(String urlString, boolean cacheEnabled) throws IOException {
-        return mImporter.loadModel(urlString, GVRImportSettings.getRecommendedSettings(), cacheEnabled, null);
-    }
-
-    
-    /**
-     * Simple, high-level method to load a scene object {@link GVRModelSceneObject} from
-     * a 3D model from a URL.
-     *
-     * @param urlString
-     *            A URL string pointing to where the model file is located.
-     *
-     * @param settings
-     *            Additional import {@link GVRImportSettings settings}
-     *
-     * @return A {@link GVRModelSceneObject} that contains the meshes with textures and bones
-     * and animations.
-     *
-     * @throws IOException
-     *             File does not exist or cannot be read
-     * @deprecated use GVRAssetLoader.loadModel instead
-     *
-     */
-    public GVRSceneObject loadModelFromURL(String urlString, EnumSet<GVRImportSettings> settings) throws IOException {
-        return mImporter.loadModel(urlString, settings, true, null);
-    }
-
-    /**
-     * Retrieves the particular index mesh for the given node.
-     * 
-     * @return The mesh, encapsulated as a {@link GVRMesh}.
-     * @deprecated use GVRContext.loadModel instead
-     */
-    GVRMesh getNodeMesh(GVRAssimpImporter assimpImporter, String nodeName, int meshIndex) {
-        return assimpImporter.getNodeMesh(nodeName, meshIndex);
-    }
 
     /**
      * Creates a quad consisting of two triangles, with the specified width and
@@ -859,283 +298,6 @@ public abstract class GVRContext implements IEventReceiver {
     }
 
     /**
-     * Loads file placed in the assets folder, as a {@link Bitmap}.
-     * 
-     * <p>
-     * Note that this method may take hundreds of milliseconds to return: unless
-     * the bitmap is quite tiny, you probably don't want to call this directly
-     * from your {@link GVRMain#onStep() onStep()} callback as that is called
-     * once per frame, and a long call will cause you to miss frames.
-     * 
-     * <p>
-     * Note also that this method does no scaling, and will return a full-size
-     * {@link Bitmap}. Loading (say) an unscaled photograph may abort your app:
-     * Use pre-scaled images, or {@link BitmapFactory} methods which give you
-     * more control over the image size.
-     * 
-     * @param fileName
-     *            The name of a file, relative to the assets directory. The
-     *            assets directory may contain an arbitrarily complex tree of
-     *            subdirectories; the file name can specify any location in or
-     *            under the assets directory.
-     * @return The file as a bitmap, or {@code null} if file path does not exist
-     *         or the file can not be decoded into a Bitmap.
-     */
-    public Bitmap loadBitmap(String fileName) {
-        if (fileName == null) {
-            throw new IllegalArgumentException("File name should not be null.");
-        }
-        InputStream stream = null;
-        Bitmap bitmap = null;
-        try {
-            try {
-                stream = mContext.getAssets().open(fileName);
-                return bitmap = BitmapFactory.decodeStream(stream);
-            } finally {
-                if (stream != null) {
-                    stream.close();
-                }
-            }
-        } catch (IOException e) {
-            e.printStackTrace();
-            // Don't discard a valid Bitmap because of an IO error closing the
-            // file!
-            return bitmap;
-        }
-    }
-
-    /**
-     * Loads file placed in the assets folder, as a {@link GVRBitmapTexture}.
-     * 
-     * <p>
-     * Note that this method may take hundreds of milliseconds to return: unless
-     * the texture is quite tiny, you probably don't want to call this directly
-     * from your {@link GVRMain#onStep() onStep()} callback as that is called
-     * once per frame, and a long call will cause you to miss frames. For large
-     * images, you should use either
-     * {@link #loadBitmapTexture(GVRAndroidResource.BitmapTextureCallback, GVRAndroidResource)
-     * loadBitmapTexture()} (faster) or
-     * {@link #loadCompressedTexture(GVRAndroidResource.CompressedTextureCallback, GVRAndroidResource)}
-     * (fastest <em>and</em> least memory pressure).
-     * 
-     * <p>
-     * Note also that this method does no scaling, and will return a full-size
-     * {@link Bitmap}. Loading (say) an unscaled photograph may abort your app:
-     * Use
-     * <ul>
-     * <li>Pre-scaled images
-     * <li>{@link BitmapFactory} methods which give you more control over the
-     * image size, or
-     * <li>
-     * {@link #loadTexture(GVRAndroidResource)} or
-     * {@link #loadBitmapTexture(GVRAndroidResource.BitmapTextureCallback, GVRAndroidResource)}
-     * which automatically scale large images to fit the GPU's restrictions and
-     * to avoid {@linkplain OutOfMemoryError out of memory errors.}
-     * </ul>
-     * 
-     * @param fileName
-     *            The name of a file, relative to the assets directory. The
-     *            assets directory may contain an arbitrarily complex tree of
-     *            sub-directories; the file name can specify any location in or
-     *            under the assets directory.
-     * @return The file as a texture, or {@code null} if file path does not
-     *         exist or the file can not be decoded into a Bitmap.
-     * 
-     * @deprecated We will remove this uncached, blocking function during Q3 of
-     *             2015. We suggest that you switch to
-     *             {@link #loadTexture(GVRAndroidResource)}
-     * 
-     */
-    public GVRBitmapTexture loadTexture(String fileName) {
-        return loadTexture(fileName, getAssetLoader().getDefaultTextureParameters());
-    }
-
-    /**
-     * Loads file placed in the assets folder, as a {@link GVRBitmapTexture}
-     * with the user provided texture parameters.
-     * 
-     * <p>
-     * Note that this method may take hundreds of milliseconds to return: unless
-     * the texture is quite tiny, you probably don't want to call this directly
-     * from your {@link GVRMain#onStep() onStep()} callback as that is called
-     * once per frame, and a long call will cause you to miss frames. For large
-     * images, you should use either
-     * {@link #loadBitmapTexture(GVRAndroidResource.BitmapTextureCallback, GVRAndroidResource)
-     * loadBitmapTexture()} (faster) or
-     * {@link #loadCompressedTexture(GVRAndroidResource.CompressedTextureCallback, GVRAndroidResource)}
-     * (fastest <em>and</em> least memory pressure).
-     * 
-     * <p>
-     * Note also that this method does no scaling, and will return a full-size
-     * {@link Bitmap}. Loading (say) an unscaled photograph may abort your app:
-     * Use
-     * <ul>
-     * <li>Pre-scaled images
-     * <li>{@link BitmapFactory} methods which give you more control over the
-     * image size, or
-     * <li>
-     * {@link #loadTexture(GVRAndroidResource)} or
-     * {@link #loadBitmapTexture(GVRAndroidResource.BitmapTextureCallback, GVRAndroidResource)}
-     * which automatically scale large images to fit the GPU's restrictions and
-     * to avoid {@linkplain OutOfMemoryError out of memory errors.}
-     * </ul>
-     * 
-     * @param fileName
-     *            The name of a file, relative to the assets directory. The
-     *            assets directory may contain an arbitrarily complex tree of
-     *            sub-directories; the file name can specify any location in or
-     *            under the assets directory.
-     * @param textureParameters
-     *            The texture parameter object which has all the values that
-     *            were provided by the user for texture enhancement. The
-     *            {@link GVRTextureParameters} class has methods to set all the
-     *            texture filters and wrap states.
-     * @return The file as a texture, or {@code null} if file path does not
-     *         exist or the file can not be decoded into a Bitmap.
-     * 
-     * @deprecated We will remove this uncached, blocking function during Q3 of
-     *             2015. We suggest that you switch to
-     *             {@link #loadTexture(GVRAndroidResource)}
-     * @deprecated use GVRAssetLoader.loadTexture instead
-     */
-    public GVRBitmapTexture loadTexture(String fileName,
-            GVRTextureParameters textureParameters) {
-        try {
-            GVRAndroidResource resource = new GVRAndroidResource(this, fileName);
-            return (GVRBitmapTexture) mImporter.loadTexture(resource, textureParameters);
-        }
-        catch (IOException ex) {
-            // The IOException will never be thrown. GVRAndroidResource says
-            // it throws an exception but it really doesn't
-            return null;
-        }
-    }
-
-    /**
-     * Loads file placed in the assets folder, as a {@link GVRBitmapTexture}.
-     * 
-     * <p>
-     * Note that this method may take hundreds of milliseconds to return: unless
-     * the texture is quite tiny, you probably don't want to call this directly
-     * from your {@link GVRMain#onStep() onStep()} callback as that is called
-     * once per frame, and a long call will cause you to miss frames. For large
-     * images, you should use either
-     * {@link #loadBitmapTexture(GVRAndroidResource.BitmapTextureCallback, GVRAndroidResource)
-     * loadBitmapTexture()} (faster) or
-     * {@link #loadCompressedTexture(GVRAndroidResource.CompressedTextureCallback, GVRAndroidResource)}
-     * (fastest <em>and</em> least memory pressure).
-     * 
-     * <p>
-     * This method automatically scales large images to fit the GPU's
-     * restrictions and to avoid {@linkplain OutOfMemoryError out of memory
-     * errors.}
-     * 
-     * @param resource
-     *            Basically, a stream containing a bitmap texture. The
-     *            {@link GVRAndroidResource} class has six constructors to
-     *            handle a wide variety of Android resource types. Taking a
-     *            {@code GVRAndroidResource} here eliminates six overloads.
-     * @return The file as a texture, or {@code null} if the file can not be
-     *         decoded into a Bitmap.
-     * 
-     * @since 1.6.5
-     * @deprecated use GVRAssetLoader.loadTexture instead
-     */
-    public GVRTexture loadTexture(GVRAndroidResource resource) {
-        return mImporter.loadTexture(resource, getAssetLoader().getDefaultTextureParameters());
-    }
-
-    /**
-     * Loads file placed in the assets folder, as a {@link GVRBitmapTexture}
-     * with the user provided texture parameters.
-     * 
-     * <p>
-     * Note that this method may take hundreds of milliseconds to return: unless
-     * the texture is quite tiny, you probably don't want to call this directly
-     * from your {@link GVRMain#onStep() onStep()} callback as that is called
-     * once per frame, and a long call will cause you to miss frames. For large
-     * images, you should use either
-     * {@link #loadBitmapTexture(GVRAndroidResource.BitmapTextureCallback, GVRAndroidResource)
-     * loadBitmapTexture()} (faster) or
-     * {@link #loadCompressedTexture(GVRAndroidResource.CompressedTextureCallback, GVRAndroidResource)}
-     * (fastest <em>and</em> least memory pressure).
-     * 
-     * <p>
-     * This method automatically scales large images to fit the GPU's
-     * restrictions and to avoid {@linkplain OutOfMemoryError out of memory
-     * errors.}
-     * 
-     * @param resource
-     *            Basically, a stream containing a bitmap texture. The
-     *            {@link GVRAndroidResource} class has six constructors to
-     *            handle a wide variety of Android resource types. Taking a
-     *            {@code GVRAndroidResource} here eliminates six overloads.
-     * @param textureParameters
-     *            The texture parameter object which has all the values that
-     *            were provided by the user for texture enhancement. The
-     *            {@link GVRTextureParameters} class has methods to set all the
-     *            texture filters and wrap states.
-     * @return The file as a texture, or {@code null} if the file can not be
-     *         decoded into a Bitmap.
-     * @deprecated use GVRAssetLoader.loadTexture instead
-     */
-    public GVRTexture loadTexture(GVRAndroidResource resource,
-            GVRTextureParameters textureParameters) {
-        return mImporter.loadTexture(resource, textureParameters);
-    }
-
-    /**
-     * Loads a cube map texture synchronously.
-     * 
-     * <p>
-     * Note that this method may take hundreds of milliseconds to return: unless
-     * the cube map is quite tiny, you probably don't want to call this directly
-     * from your {@link GVRMain#onStep() onStep()} callback as that is called
-     * once per frame, and a long call will cause you to miss frames.
-     * 
-     * @param resourceArray
-     *            An array containing six resources for six bitmaps. The order
-     *            of the bitmaps is important to the correctness of the cube map
-     *            texture. The six bitmaps should correspond to +x, -x, +y, -y,
-     *            +z, and -z faces of the cube map texture respectively.
-     * @return The cube map texture, or {@code null} if the length of
-     *         rsourceArray is not 6.
-     * 
-     * @since 1.6.9
-     */
-    /*
-     * TODO Deprecate, and replace with an overload that takes a single
-     * GVRAndroidResource which specifies a zip file ... and caches result
-     */
-    public GVRCubemapTexture loadCubemapTexture(
-            GVRAndroidResource[] resourceArray) {
-        return loadCubemapTexture(resourceArray, getAssetLoader().getDefaultTextureParameters());
-    }
-
-    // Texture parameters
-    public GVRCubemapTexture loadCubemapTexture(
-            GVRAndroidResource[] resourceArray,
-            GVRTextureParameters textureParameters) {
-
-        if (resourceArray.length != 6) {
-            return null;
-        }
-
-        Bitmap[] bitmapArray = new Bitmap[6];
-        try {
-            for (int i = 0; i < 6; i++) {
-                bitmapArray[i] = GVRAsynchronousResourceLoader.decodeStream(
-                        resourceArray[i].getStream(), false);
-                resourceArray[i].closeStream();
-            }
-        }
-        catch (IOException ex) {
-            return null;
-        }
-        return new GVRCubemapTexture(this, bitmapArray, textureParameters);
-    }
-
-    /**
      * Throws an exception if the current thread is not a GL thread.
      * 
      * @since 1.6.5
@@ -1159,1171 +321,6 @@ public abstract class GVRContext implements IEventReceiver {
 
     public boolean isCurrentThreadGLThread() {
         return Thread.currentThread().getId() == mGLThreadID;
-    }
-
-    /**
-     * Load a bitmap, asynchronously, with a default priority.
-     * 
-     * Because it is asynchronous, this method <em>is</em> a bit harder to use
-     * than {@link #loadTexture(String)}, but it moves a large amount of work
-     * (in {@link BitmapFactory#decodeStream(InputStream)} from the GL thread to
-     * a background thread. Since you <em>can</em> create a
-     * {@link GVRSceneObject} without a mesh and texture - and set them later -
-     * using the asynchronous API can improve startup speed and/or reduce frame
-     * misses (where an {@link GVRMain#onStep() onStep()} takes too long).
-     * This API may also use less RAM than {@link #loadTexture(String)}.
-     * 
-     * <p>
-     * This API will 'consolidate' requests: If you request a texture like
-     * {@code R.raw.wood_grain} and then - before it has loaded - issue another
-     * request for {@code R.raw.wood_grain}, GVRF will only read the bitmap file
-     * once; only create a single {@link GVRTexture}; and then call both
-     * callbacks, passing each the same texture.
-     * 
-     * <p>
-     * Please be aware that {@link BitmapFactory#decodeStream(InputStream)} is a
-     * comparatively expensive operation: it can take hundreds of milliseconds
-     * and use several megabytes of temporary RAM. GVRF includes a throttler to
-     * keep the total load manageable - but
-     * {@link #loadCompressedTexture(GVRAndroidResource.CompressedTextureCallback, GVRAndroidResource)}
-     * is <em>much</em> faster and lighter-weight: that API simply loads the
-     * compressed texture into a small amount RAM (which doesn't take very long)
-     * and does some simple parsing to figure out the parameters to pass
-     * {@code glCompressedTexImage2D()}. The GL hardware does the decoding much
-     * faster than Android's {@link BitmapFactory}!
-     * 
-     * <p>
-     * TODO Take a boolean parameter that controls mipmap generation?
-     * 
-     * @since 1.6.1
-     * 
-     * @param callback
-     *            Before loading, GVRF may call
-     *            {@link GVRAndroidResource.BitmapTextureCallback#stillWanted(GVRAndroidResource)
-     *            stillWanted()} several times (on a background thread) to give
-     *            you a chance to abort a 'stale' load.
-     * 
-     *            Successful loads will call
-     *            {@link GVRAndroidResource.Callback#loaded(GVRHybridObject, GVRAndroidResource)
-     *            loaded()} on the GL thread;
-     * 
-     *            any errors will call
-     *            {@link GVRAndroidResource.BitmapTextureCallback#failed(Throwable, GVRAndroidResource)
-     *            failed()}, with no promises about threading.
-     * 
-     *            <p>
-     *            This method uses a throttler to avoid overloading the system.
-     *            If the throttler has threads available, it will run this
-     *            request immediately. Otherwise, it will enqueue the request,
-     *            and call
-     *            {@link GVRAndroidResource.BitmapTextureCallback#stillWanted(GVRAndroidResource)
-     *            stillWanted()} at least once (on a background thread) to give
-     *            you a chance to abort a 'stale' load.
-     * @param resource
-     *            Basically, a stream containing a bitmapped image. The
-     *            {@link GVRAndroidResource} class has six constructors to
-     *            handle a wide variety of Android resource types. Taking a
-     *            {@code GVRAndroidResource} here eliminates six overloads.
-     * 
-     * @throws IllegalArgumentException
-     *             If you 'abuse' request consolidation by passing the same
-     *             {@link GVRAndroidResource} descriptor to multiple load calls.
-     *             <p>
-     *             It's fairly common for multiple scene objects to use the same
-     *             texture or the same mesh. Thus, if you try to load, say,
-     *             {@code R.raw.whatever} while you already have a pending
-     *             request for {@code R.raw.whatever}, it will only be loaded
-     *             once; the same resource will be used to satisfy both (all)
-     *             requests. This "consolidation" uses
-     *             {@link GVRAndroidResource#equals(Object)}, <em>not</em>
-     *             {@code ==} (aka "reference equality"): The problem with using
-     *             the same resource descriptor is that if requests can't be
-     *             consolidated (because the later one(s) came in after the
-     *             earlier one(s) had already completed) the resource will be
-     *             reloaded ... but the original descriptor will have been
-     *             closed.
-     * @deprecated use GVRAssetloader.loadTexture instead
-     */
-    public void loadBitmapTexture(BitmapTextureCallback callback,
-            GVRAndroidResource resource) {
-        mImporter.loadTexture(resource, callback, null, DEFAULT_PRIORITY, GVRCompressedTexture.BALANCED);
-    }
-
-    /**
-     * Load a bitmap, asynchronously, with an explicit priority.
-     * 
-     * Because it is asynchronous, this method <em>is</em> a bit harder to use
-     * than {@link #loadTexture(String)}, but it moves a large amount of work
-     * (in {@link BitmapFactory#decodeStream(InputStream)} from the GL thread to
-     * a background thread. Since you <em>can</em> create a
-     * {@link GVRSceneObject} without a mesh and texture - and set them later -
-     * using the asynchronous API can improve startup speed and/or reduce frame
-     * misses, where an {@link GVRMain#onStep() onStep()} takes too long.
-     * 
-     * <p>
-     * This API will 'consolidate' requests: If you request a texture like
-     * {@code R.raw.wood_grain} and then - before it has loaded - issue another
-     * request for {@code R.raw.wood_grain}, GVRF will only read the bitmap file
-     * once; only create a single {@link GVRTexture}; and then call both
-     * callbacks, passing each the same texture.
-     * 
-     * <p>
-     * Please be aware that {@link BitmapFactory#decodeStream(InputStream)} is a
-     * comparatively expensive operation: it can take hundreds of milliseconds
-     * and use several megabytes of temporary RAM. GVRF includes a throttler to
-     * keep the total load manageable - but
-     * {@link #loadCompressedTexture(GVRAndroidResource.CompressedTextureCallback, GVRAndroidResource)}
-     * is <em>much</em> faster and lighter-weight: that API simply loads the
-     * compressed texture into a small amount RAM (which doesn't take very long)
-     * and does some simple parsing to figure out the parameters to pass
-     * {@code glCompressedTexImage2D()}. The GL hardware does the decoding much
-     * faster than Android's {@link BitmapFactory}!
-     * 
-     * @since 1.6.1
-     * 
-     * @param callback
-     *            Before loading, GVRF may call
-     *            {@link GVRAndroidResource.BitmapTextureCallback#stillWanted(GVRAndroidResource)
-     *            stillWanted()} several times (on a background thread) to give
-     *            you a chance to abort a 'stale' load.
-     * 
-     *            Successful loads will call
-     *            {@link GVRAndroidResource.Callback#loaded(GVRHybridObject, GVRAndroidResource)
-     *            loaded()} on the GL thread;
-     * 
-     *            any errors will call
-     *            {@link GVRAndroidResource.BitmapTextureCallback#failed(Throwable, GVRAndroidResource)
-     *            failed()}, with no promises about threading.
-     * 
-     *            <p>
-     *            This method uses a throttler to avoid overloading the system.
-     *            If the throttler has threads available, it will run this
-     *            request immediately. Otherwise, it will enqueue the request,
-     *            and call
-     *            {@link GVRAndroidResource.BitmapTextureCallback#stillWanted(GVRAndroidResource)
-     *            stillWanted()} at least once (on a background thread) to give
-     *            you a chance to abort a 'stale' load.
-     * @param resource
-     *            Basically, a stream containing a bitmapped image. The
-     *            {@link GVRAndroidResource} class has six constructors to
-     *            handle a wide variety of Android resource types. Taking a
-     *            {@code GVRAndroidResource} here eliminates six overloads.
-     * @param priority
-     *            This request's priority. Please see the notes on asynchronous
-     *            priorities in the <a href="package-summary.html#async">package
-     *            description</a>.
-     * 
-     * @throws IllegalArgumentException
-     *             If {@code priority} {@literal <} {@link #LOWEST_PRIORITY} or
-     *             {@literal >} {@link #HIGHEST_PRIORITY}, or either of the
-     *             other parameters is {@code null} - or if you 'abuse' request
-     *             consolidation by passing the same {@link GVRAndroidResource}
-     *             descriptor to multiple load calls.
-     *             <p>
-     *             It's fairly common for multiple scene objects to use the same
-     *             texture or the same mesh. Thus, if you try to load, say,
-     *             {@code R.raw.whatever} while you already have a pending
-     *             request for {@code R.raw.whatever}, it will only be loaded
-     *             once; the same resource will be used to satisfy both (all)
-     *             requests. This "consolidation" uses
-     *             {@link GVRAndroidResource#equals(Object)}, <em>not</em>
-     *             {@code ==} (aka "reference equality"): The problem with using
-     *             the same resource descriptor is that if requests can't be
-     *             consolidated (because the later one(s) came in after the
-     *             earlier one(s) had already completed) the resource will be
-     *             reloaded ... but the original descriptor will have been
-     *             closed.
-     *  @deprecated use GVRAssetloader.loadTexture instead
-     */
-    public void loadBitmapTexture(TextureCallback callback,
-            GVRAndroidResource resource, int priority)
-            throws IllegalArgumentException {
-        mImporter.loadTexture(resource, callback, null, priority, GVRCompressedTexture.BALANCED);
-    }
-
-    /**
-     * Load a compressed texture, asynchronously.
-     * 
-     * GVRF currently supports ASTC, ETC2, and KTX formats: applications can add
-     * new formats by implementing {@link GVRCompressedTextureLoader}.
-     * 
-     * <p>
-     * This method uses the fastest possible rendering. To specify higher
-     * quality (but slower) rendering, you can use the
-     * {@link #loadCompressedTexture(GVRAndroidResource.CompressedTextureCallback, GVRAndroidResource, int)}
-     * overload.
-     * 
-     * @since 1.6.1
-     * 
-     * @param callback
-     *            Successful loads will call
-     *            {@link GVRAndroidResource.Callback#loaded(GVRHybridObject, GVRAndroidResource)
-     *            loaded()} on the GL thread; any errors will call
-     *            {@link GVRAndroidResource.CompressedTextureCallback#failed(Throwable, GVRAndroidResource)
-     *            failed()}, with no promises about threading.
-     * @param resource
-     *            Basically, a stream containing a compressed texture. The
-     *            {@link GVRAndroidResource} class has six constructors to
-     *            handle a wide variety of Android resource types. Taking a
-     *            {@code GVRAndroidResource} here eliminates six overloads.
-     * 
-     * @throws IllegalArgumentException
-     *             If you 'abuse' request consolidation by passing the same
-     *             {@link GVRAndroidResource} descriptor to multiple load calls.
-     *             <p>
-     *             It's fairly common for multiple scene objects to use the same
-     *             texture or the same mesh. Thus, if you try to load, say,
-     *             {@code R.raw.whatever} while you already have a pending
-     *             request for {@code R.raw.whatever}, it will only be loaded
-     *             once; the same resource will be used to satisfy both (all)
-     *             requests. This "consolidation" uses
-     *             {@link GVRAndroidResource#equals(Object)}, <em>not</em>
-     *             {@code ==} (aka "reference equality"): The problem with using
-     *             the same resource descriptor is that if requests can't be
-     *             consolidated (because the later one(s) came in after the
-     *             earlier one(s) had already completed) the resource will be
-     *             reloaded ... but the original descriptor will have been
-     *             closed.
-     *  @deprecated use GVRAssetloader.loadTexture instead
-     */
-    public void loadCompressedTexture(CompressedTextureCallback callback,
-            GVRAndroidResource resource) {
-        mImporter.loadTexture(resource, callback);
-    }
-
-    /**
-     * Load a compressed texture, asynchronously.
-     * 
-     * GVRF currently supports ASTC, ETC2, and KTX formats: applications can add
-     * new formats by implementing {@link GVRCompressedTextureLoader}.
-     * 
-     * @since 1.6.1
-     * 
-     * @param callback
-     *            Successful loads will call
-     *            {@link GVRAndroidResource.Callback#loaded(GVRHybridObject, GVRAndroidResource)}
-     *            on the GL thread; any errors will call
-     *            {@link GVRAndroidResource.CompressedTextureCallback#failed(Throwable, GVRAndroidResource)}
-     *            , with no promises about threading.
-     * @param resource
-     *            Basically, a stream containing a compressed texture. The
-     *            {@link GVRAndroidResource} class has six constructors to
-     *            handle a wide variety of Android resource types. Taking a
-     *            {@code GVRAndroidResource} here eliminates six overloads.
-     * @param quality
-     *            Speed/quality tradeoff: should be one of
-     *            {@link GVRCompressedTexture#SPEED},
-     *            {@link GVRCompressedTexture#BALANCED}, or
-     *            {@link GVRCompressedTexture#QUALITY}, but other values are
-     *            'clamped' to one of the recognized values.
-     * 
-     * @throws IllegalArgumentException
-     *             If you 'abuse' request consolidation by passing the same
-     *             {@link GVRAndroidResource} descriptor to multiple load calls.
-     *             <p>
-     *             It's fairly common for multiple scene objects to use the same
-     *             texture or the same mesh. Thus, if you try to load, say,
-     *             {@code R.raw.whatever} while you already have a pending
-     *             request for {@code R.raw.whatever}, it will only be loaded
-     *             once; the same resource will be used to satisfy both (all)
-     *             requests. This "consolidation" uses
-     *             {@link GVRAndroidResource#equals(Object)}, <em>not</em>
-     *             {@code ==} (aka "reference equality"): The problem with using
-     *             the same resource descriptor is that if requests can't be
-     *             consolidated (because the later one(s) came in after the
-     *             earlier one(s) had already completed) the resource will be
-     *             reloaded ... but the original descriptor will have been
-     *             closed.
-     *  @deprecated use GVRAssetloader.loadTexture instead
-     */
-    public void loadCompressedTexture(CompressedTextureCallback callback,
-            GVRAndroidResource resource, int quality) {
-        mImporter.loadTexture(resource, callback, null, GVRAssetLoader.DEFAULT_PRIORITY, quality);
-    }
-
-    /**
-     * A simplified, low-level method that loads a texture asynchronously,
-     * without making you specify
-     * {@link #loadBitmapTexture(GVRAndroidResource.BitmapTextureCallback, GVRAndroidResource)
-     * loadBitmapTexture()} or
-     * {@link #loadCompressedTexture(GVRAndroidResource.CompressedTextureCallback, GVRAndroidResource)
-     * loadCompressedTexture()}.
-     * 
-     * This method can detect whether the resource file holds a compressed
-     * texture (GVRF currently supports ASTC, ETC2, and KTX formats:
-     * applications can add new formats by implementing
-     * {@link GVRCompressedTextureLoader}): if the file is not a compressed
-     * texture, it is loaded as a normal, bitmapped texture. This format
-     * detection adds very little to the cost of loading even a compressed
-     * texture, and it makes your life a lot easier: you can replace, say,
-     * {@code res/raw/resource.png} with {@code res/raw/resource.etc2} without
-     * having to change any code.
-     * 
-     * <p>
-     * This method uses a default priority and a default render quality: Use
-     * {@link #loadTexture(GVRAndroidResource.TextureCallback, GVRAndroidResource, int)}
-     * to specify an explicit priority, and
-     * {@link #loadTexture(GVRAndroidResource.TextureCallback, GVRAndroidResource, int, int)}
-     * to specify an explicit quality.
-     * 
-     * <p>
-     * We will continue to support the {@code loadBitmapTexture()} and
-     * {@code loadCompressedTexture()} APIs for at least a little while: We
-     * haven't yet decided whether to deprecate them or not.
-     * 
-     * @param callback
-     *            Before loading, GVRF may call
-     *            {@link GVRAndroidResource.TextureCallback#stillWanted(GVRAndroidResource)
-     *            stillWanted()} several times (on a background thread) to give
-     *            you a chance to abort a 'stale' load.
-     * 
-     *            Successful loads will call
-     *            {@link GVRAndroidResource.Callback#loaded(GVRHybridObject, GVRAndroidResource)
-     *            loaded()} on the GL thread;
-     * 
-     *            any errors will call
-     *            {@link GVRAndroidResource.TextureCallback#failed(Throwable, GVRAndroidResource)
-     *            failed()}, with no promises about threading.
-     * 
-     *            <p>
-     *            This method uses a throttler to avoid overloading the system.
-     *            If the throttler has threads available, it will run this
-     *            request immediately. Otherwise, it will enqueue the request,
-     *            and call
-     *            {@link GVRAndroidResource.TextureCallback#stillWanted(GVRAndroidResource)
-     *            stillWanted()} at least once (on a background thread) to give
-     *            you a chance to abort a 'stale' load.
-     * 
-     *            <p>
-     *            Use {@link #loadFutureTexture(GVRAndroidResource)} to avoid
-     *            having to implement a callback.
-     * @param resource
-     *            Basically, a stream containing a texture file. The
-     *            {@link GVRAndroidResource} class has six constructors to
-     *            handle a wide variety of Android resource types. Taking a
-     *            {@code GVRAndroidResource} here eliminates six overloads.
-     * 
-     * @since 1.6.7
-     * 
-     * @throws IllegalArgumentException
-     *             If you 'abuse' request consolidation by passing the same
-     *             {@link GVRAndroidResource} descriptor to multiple load calls.
-     *             <p>
-     *             It's fairly common for multiple scene objects to use the same
-     *             texture or the same mesh. Thus, if you try to load, say,
-     *             {@code R.raw.whatever} while you already have a pending
-     *             request for {@code R.raw.whatever}, it will only be loaded
-     *             once; the same resource will be used to satisfy both (all)
-     *             requests. This "consolidation" uses
-     *             {@link GVRAndroidResource#equals(Object)}, <em>not</em>
-     *             {@code ==} (aka "reference equality"): The problem with using
-     *             the same resource descriptor is that if requests can't be
-     *             consolidated (because the later one(s) came in after the
-     *             earlier one(s) had already completed) the resource will be
-     *             reloaded ... but the original descriptor will have been
-     *             closed.
-     *  @deprecated use GVRAssetloader.loadTexture instead
-     */
-    public void loadTexture(TextureCallback callback,
-            GVRAndroidResource resource) {
-        mImporter.loadTexture(resource, callback, null, GVRAssetLoader.DEFAULT_PRIORITY, GVRCompressedTexture.BALANCED);
-    }
-
-    /**
-     * A simplified, low-level method that loads a texture asynchronously,
-     * without making you specify
-     * {@link #loadBitmapTexture(GVRAndroidResource.BitmapTextureCallback, GVRAndroidResource)
-     * loadBitmapTexture()} or
-     * {@link #loadCompressedTexture(GVRAndroidResource.CompressedTextureCallback, GVRAndroidResource)
-     * loadCompressedTexture()}.
-     * 
-     * This method can detect whether the resource file holds a compressed
-     * texture (GVRF currently supports ASTC, ETC2, and KTX formats:
-     * applications can add new formats by implementing
-     * {@link GVRCompressedTextureLoader}): if the file is not a compressed
-     * texture, it is loaded as a normal, bitmapped texture. This format
-     * detection adds very little to the cost of loading even a compressed
-     * texture, and it makes your life a lot easier: you can replace, say,
-     * {@code res/raw/resource.png} with {@code res/raw/resource.etc2} without
-     * having to change any code.
-     * 
-     * <p>
-     * This method uses a default render quality: Use
-     * {@link #loadTexture(GVRAndroidResource.TextureCallback, GVRAndroidResource, int, int)}
-     * to specify an explicit quality.
-     * 
-     * <p>
-     * We will continue to support the {@code loadBitmapTexture()} and
-     * {@code loadCompressedTexture()} APIs for at least a little while: We
-     * haven't yet decided whether to deprecate them or not.
-     * 
-     * @param callback
-     *            Before loading, GVRF may call
-     *            {@link GVRAndroidResource.TextureCallback#stillWanted(GVRAndroidResource)
-     *            stillWanted()} several times (on a background thread) to give
-     *            you a chance to abort a 'stale' load.
-     * 
-     *            Successful loads will call
-     *            {@link GVRAndroidResource.Callback#loaded(GVRHybridObject, GVRAndroidResource)
-     *            loaded()} on the GL thread;
-     * 
-     *            any errors will call
-     *            {@link GVRAndroidResource.TextureCallback#failed(Throwable, GVRAndroidResource)
-     *            failed()}, with no promises about threading.
-     * 
-     *            <p>
-     *            This method uses a throttler to avoid overloading the system.
-     *            If the throttler has threads available, it will run this
-     *            request immediately. Otherwise, it will enqueue the request,
-     *            and call
-     *            {@link GVRAndroidResource.TextureCallback#stillWanted(GVRAndroidResource)
-     *            stillWanted()} at least once (on a background thread) to give
-     *            you a chance to abort a 'stale' load.
-     * 
-     *            <p>
-     *            Use {@link #loadFutureTexture(GVRAndroidResource)} to avoid
-     *            having to implement a callback.
-     * @param resource
-     *            Basically, a stream containing a texture file. The
-     *            {@link GVRAndroidResource} class has six constructors to
-     *            handle a wide variety of Android resource types. Taking a
-     *            {@code GVRAndroidResource} here eliminates six overloads.
-     * @param priority
-     *            This request's priority. Please see the notes on asynchronous
-     *            priorities in the <a href="package-summary.html#async">package
-     *            description</a>. Also, please note priorities only apply to
-     *            uncompressed textures (standard Android bitmap files, which
-     *            can take hundreds of milliseconds to load): compressed
-     *            textures load so quickly that they are not run through the
-     *            request scheduler.
-     * 
-     * @since 1.6.7
-     * 
-     * @throws IllegalArgumentException
-     *             If you 'abuse' request consolidation by passing the same
-     *             {@link GVRAndroidResource} descriptor to multiple load calls.
-     *             <p>
-     *             It's fairly common for multiple scene objects to use the same
-     *             texture or the same mesh. Thus, if you try to load, say,
-     *             {@code R.raw.whatever} while you already have a pending
-     *             request for {@code R.raw.whatever}, it will only be loaded
-     *             once; the same resource will be used to satisfy both (all)
-     *             requests. This "consolidation" uses
-     *             {@link GVRAndroidResource#equals(Object)}, <em>not</em>
-     *             {@code ==} (aka "reference equality"): The problem with using
-     *             the same resource descriptor is that if requests can't be
-     *             consolidated (because the later one(s) came in after the
-     *             earlier one(s) had already completed) the resource will be
-     *             reloaded ... but the original descriptor will have been
-     *             closed.
-     *  @deprecated use GVRAssetloader.loadTexture instead
-     */
-    public void loadTexture(TextureCallback callback,
-            GVRAndroidResource resource, int priority) {
-        mImporter.loadTexture(resource, callback, null, priority, GVRCompressedTexture.BALANCED);
-    }
-
-    /**
-     * A simplified, low-level method that loads a texture asynchronously,
-     * without making you specify
-     * {@link #loadBitmapTexture(GVRAndroidResource.BitmapTextureCallback, GVRAndroidResource)
-     * loadBitmapTexture()} or
-     * {@link #loadCompressedTexture(GVRAndroidResource.CompressedTextureCallback, GVRAndroidResource)
-     * loadCompressedTexture()}.
-     * 
-     * This method can detect whether the resource file holds a compressed
-     * texture (GVRF currently supports ASTC, ETC2, and KTX formats:
-     * applications can add new formats by implementing
-     * {@link GVRCompressedTextureLoader}): if the file is not a compressed
-     * texture, it is loaded as a normal, bitmapped texture. This format
-     * detection adds very little to the cost of loading even a compressed
-     * texture, and it makes your life a lot easier: you can replace, say,
-     * {@code res/raw/resource.png} with {@code res/raw/resource.etc2} without
-     * having to change any code.
-     * 
-     * <p>
-     * We will continue to support the {@code loadBitmapTexture()} and
-     * {@code loadCompressedTexture()} APIs for at least a little while: We
-     * haven't yet decided whether to deprecate them or not.
-     * 
-     * @param callback
-     *            Before loading, GVRF may call
-     *            {@link GVRAndroidResource.TextureCallback#stillWanted(GVRAndroidResource)
-     *            stillWanted()} several times (on a background thread) to give
-     *            you a chance to abort a 'stale' load.
-     * 
-     *            Successful loads will call
-     *            {@link GVRAndroidResource.Callback#loaded(GVRHybridObject, GVRAndroidResource)
-     *            loaded()} on the GL thread;
-     * 
-     *            any errors will call
-     *            {@link GVRAndroidResource.TextureCallback#failed(Throwable, GVRAndroidResource)
-     *            failed()}, with no promises about threading.
-     * 
-     *            <p>
-     *            This method uses a throttler to avoid overloading the system.
-     *            If the throttler has threads available, it will run this
-     *            request immediately. Otherwise, it will enqueue the request,
-     *            and call
-     *            {@link GVRAndroidResource.TextureCallback#stillWanted(GVRAndroidResource)
-     *            stillWanted()} at least once (on a background thread) to give
-     *            you a chance to abort a 'stale' load.
-     * 
-     *            <p>
-     *            Use {@link #loadFutureTexture(GVRAndroidResource)} to avoid
-     *            having to implement a callback.
-     * @param resource
-     *            Basically, a stream containing a texture file. The
-     *            {@link GVRAndroidResource} class has six constructors to
-     *            handle a wide variety of Android resource types. Taking a
-     *            {@code GVRAndroidResource} here eliminates six overloads.
-     * @param priority
-     *            This request's priority. Please see the notes on asynchronous
-     *            priorities in the <a href="package-summary.html#async">package
-     *            description</a>. Also, please note priorities only apply to
-     *            uncompressed textures (standard Android bitmap files, which
-     *            can take hundreds of milliseconds to load): compressed
-     *            textures load so quickly that they are not run through the
-     *            request scheduler.
-     * @param quality
-     *            The compressed texture {@link GVRCompressedTexture#mQuality
-     *            quality} parameter: should be one of
-     *            {@link GVRCompressedTexture#SPEED},
-     *            {@link GVRCompressedTexture#BALANCED}, or
-     *            {@link GVRCompressedTexture#QUALITY}, but other values are
-     *            'clamped' to one of the recognized values. Please note that
-     *            this (currently) only applies to compressed textures; normal
-     *            {@linkplain GVRBitmapTexture bitmapped textures} don't take a
-     *            quality parameter.
-     *
-     * @since 1.6.7
-     *
-     * @throws IllegalArgumentException
-     *             If you 'abuse' request consolidation by passing the same
-     *             {@link GVRAndroidResource} descriptor to multiple load calls.
-     *             <p>
-     *             It's fairly common for multiple scene objects to use the same
-     *             texture or the same mesh. Thus, if you try to load, say,
-     *             {@code R.raw.whatever} while you already have a pending
-     *             request for {@code R.raw.whatever}, it will only be loaded
-     *             once; the same resource will be used to satisfy both (all)
-     *             requests. This "consolidation" uses
-     *             {@link GVRAndroidResource#equals(Object)}, <em>not</em>
-     *             {@code ==} (aka "reference equality"): The problem with using
-     *             the same resource descriptor is that if requests can't be
-     *             consolidated (because the later one(s) came in after the
-     *             earlier one(s) had already completed) the resource will be
-     *             reloaded ... but the original descriptor will have been
-     *             closed.
-     * @deprecated use GVRAssetloader.loadTexture instead
-     */
-    public void loadTexture(TextureCallback callback,
-            GVRAndroidResource resource, int priority, int quality) {
-        mImporter.loadTexture(resource, callback, null, priority, quality);
-    }
-
-    /**
-     * A simplified, low-level method that loads a texture asynchronously,
-     * without making you specify
-     * {@link #loadBitmapTexture(GVRAndroidResource.BitmapTextureCallback, GVRAndroidResource)
-     * loadBitmapTexture()} or
-     * {@link #loadCompressedTexture(GVRAndroidResource.CompressedTextureCallback, GVRAndroidResource)
-     * loadCompressedTexture()}.
-     * 
-     * This method can detect whether the resource file holds a compressed
-     * texture (GVRF currently supports ASTC, ETC2, and KTX formats:
-     * applications can add new formats by implementing
-     * {@link GVRCompressedTextureLoader}): if the file is not a compressed
-     * texture, it is loaded as a normal, bitmapped texture. This format
-     * detection adds very little to the cost of loading even a compressed
-     * texture, and it makes your life a lot easier: you can replace, say,
-     * {@code res/raw/resource.png} with {@code res/raw/resource.etc2} without
-     * having to change any code.
-     * 
-     * <p>
-     * This method uses a default priority and a default render quality: Use
-     * {@link #loadTexture(GVRAndroidResource.TextureCallback, GVRAndroidResource, int)}
-     * to specify an explicit priority, and
-     * {@link #loadTexture(GVRAndroidResource.TextureCallback, GVRAndroidResource, int, int)}
-     * to specify an explicit quality.
-     * 
-     * <p>
-     * We will continue to support the {@code loadBitmapTexture()} and
-     * {@code loadCompressedTexture()} APIs for at least a little while: We
-     * haven't yet decided whether to deprecate them or not.
-     * 
-     * @param callback
-     *            Before loading, GVRF may call
-     *            {@link GVRAndroidResource.TextureCallback#stillWanted(GVRAndroidResource)
-     *            stillWanted()} several times (on a background thread) to give
-     *            you a chance to abort a 'stale' load.
-     * 
-     *            Successful loads will call
-     *            {@link GVRAndroidResource.Callback#loaded(GVRHybridObject, GVRAndroidResource)
-     *            loaded()} on the GL thread;
-     * 
-     *            any errors will call
-     *            {@link GVRAndroidResource.TextureCallback#failed(Throwable, GVRAndroidResource)
-     *            failed()}, with no promises about threading.
-     * 
-     *            <p>
-     *            This method uses a throttler to avoid overloading the system.
-     *            If the throttler has threads available, it will run this
-     *            request immediately. Otherwise, it will enqueue the request,
-     *            and call
-     *            {@link GVRAndroidResource.TextureCallback#stillWanted(GVRAndroidResource)
-     *            stillWanted()} at least once (on a background thread) to give
-     *            you a chance to abort a 'stale' load.
-     * 
-     *            <p>
-     *            Use {@link #loadFutureTexture(GVRAndroidResource)} to avoid
-     *            having to implement a callback.
-     * @param resource
-     *            Basically, a stream containing a texture file. The
-     *            {@link GVRAndroidResource} class has six constructors to
-     *            handle a wide variety of Android resource types. Taking a
-     *            {@code GVRAndroidResource} here eliminates six overloads.
-     * @param textureParameters
-     *            The texture parameter object which has all the values that
-     *            were provided by the user for texture enhancement. The
-     *            {@link GVRTextureParameters} class has methods to set all the
-     *            texture filters and wrap states.
-     * 
-     * @since 2.0.2
-     * 
-     * @throws IllegalArgumentException
-     *             If you 'abuse' request consolidation by passing the same
-     *             {@link GVRAndroidResource} descriptor to multiple load calls.
-     *             <p>
-     *             It's fairly common for multiple scene objects to use the same
-     *             texture or the same mesh. Thus, if you try to load, say,
-     *             {@code R.raw.whatever} while you already have a pending
-     *             request for {@code R.raw.whatever}, it will only be loaded
-     *             once; the same resource will be used to satisfy both (all)
-     *             requests. This "consolidation" uses
-     *             {@link GVRAndroidResource#equals(Object)}, <em>not</em>
-     *             {@code ==} (aka "reference equality"): The problem with using
-     *             the same resource descriptor is that if requests can't be
-     *             consolidated (because the later one(s) came in after the
-     *             earlier one(s) had already completed) the resource will be
-     *             reloaded ... but the original descriptor will have been
-     *             closed.
-     */
-    public void loadTexture(TextureCallback callback,
-            GVRAndroidResource resource, GVRTextureParameters textureParameters) {
-        mImporter.loadTexture(resource, callback, textureParameters, GVRAssetLoader.DEFAULT_PRIORITY, GVRCompressedTexture.BALANCED);
-    }
-
-    /**
-     * A simplified, low-level method that loads a texture asynchronously,
-     * without making you specify
-     * {@link #loadBitmapTexture(GVRAndroidResource.BitmapTextureCallback, GVRAndroidResource)
-     * loadBitmapTexture()} or
-     * {@link #loadCompressedTexture(GVRAndroidResource.CompressedTextureCallback, GVRAndroidResource)
-     * loadCompressedTexture()}.
-     * 
-     * This method can detect whether the resource file holds a compressed
-     * texture (GVRF currently supports ASTC, ETC2, and KTX formats:
-     * applications can add new formats by implementing
-     * {@link GVRCompressedTextureLoader}): if the file is not a compressed
-     * texture, it is loaded as a normal, bitmapped texture. This format
-     * detection adds very little to the cost of loading even a compressed
-     * texture, and it makes your life a lot easier: you can replace, say,
-     * {@code res/raw/resource.png} with {@code res/raw/resource.etc2} without
-     * having to change any code.
-     * 
-     * <p>
-     * This method uses a default render quality: Use
-     * {@link #loadTexture(GVRAndroidResource.TextureCallback, GVRAndroidResource, int, int)}
-     * to specify an explicit quality.
-     * 
-     * <p>
-     * We will continue to support the {@code loadBitmapTexture()} and
-     * {@code loadCompressedTexture()} APIs for at least a little while: We
-     * haven't yet decided whether to deprecate them or not.
-     * 
-     * @param callback
-     *            Before loading, GVRF may call
-     *            {@link GVRAndroidResource.TextureCallback#stillWanted(GVRAndroidResource)
-     *            stillWanted()} several times (on a background thread) to give
-     *            you a chance to abort a 'stale' load.
-     * 
-     *            Successful loads will call
-     *            {@link GVRAndroidResource.Callback#loaded(GVRHybridObject, GVRAndroidResource)
-     *            loaded()} on the GL thread;
-     * 
-     *            any errors will call
-     *            {@link GVRAndroidResource.TextureCallback#failed(Throwable, GVRAndroidResource)
-     *            failed()}, with no promises about threading.
-     * 
-     *            <p>
-     *            This method uses a throttler to avoid overloading the system.
-     *            If the throttler has threads available, it will run this
-     *            request immediately. Otherwise, it will enqueue the request,
-     *            and call
-     *            {@link GVRAndroidResource.TextureCallback#stillWanted(GVRAndroidResource)
-     *            stillWanted()} at least once (on a background thread) to give
-     *            you a chance to abort a 'stale' load.
-     * 
-     *            <p>
-     *            Use {@link #loadFutureTexture(GVRAndroidResource)} to avoid
-     *            having to implement a callback.
-     * @param resource
-     *            Basically, a stream containing a texture file. The
-     *            {@link GVRAndroidResource} class has six constructors to
-     *            handle a wide variety of Android resource types. Taking a
-     *            {@code GVRAndroidResource} here eliminates six overloads.
-     * @param textureParameters
-     *            The texture parameter object which has all the values that
-     *            were provided by the user for texture enhancement. The
-     *            {@link GVRTextureParameters} class has methods to set all the
-     *            texture filters and wrap states.
-     * @param priority
-     *            This request's priority. Please see the notes on asynchronous
-     *            priorities in the <a href="package-summary.html#async">package
-     *            description</a>. Also, please note priorities only apply to
-     *            uncompressed textures (standard Android bitmap files, which
-     *            can take hundreds of milliseconds to load): compressed
-     *            textures load so quickly that they are not run through the
-     *            request scheduler.
-     * @since 2.0.2
-     * 
-     * @throws IllegalArgumentException
-     *             If you 'abuse' request consolidation by passing the same
-     *             {@link GVRAndroidResource} descriptor to multiple load calls.
-     *             <p>
-     *             It's fairly common for multiple scene objects to use the same
-     *             texture or the same mesh. Thus, if you try to load, say,
-     *             {@code R.raw.whatever} while you already have a pending
-     *             request for {@code R.raw.whatever}, it will only be loaded
-     *             once; the same resource will be used to satisfy both (all)
-     *             requests. This "consolidation" uses
-     *             {@link GVRAndroidResource#equals(Object)}, <em>not</em>
-     *             {@code ==} (aka "reference equality"): The problem with using
-     *             the same resource descriptor is that if requests can't be
-     *             consolidated (because the later one(s) came in after the
-     *             earlier one(s) had already completed) the resource will be
-     *             reloaded ... but the original descriptor will have been
-     *             closed.
-     */
-    public void loadTexture(TextureCallback callback,
-            GVRAndroidResource resource, GVRTextureParameters textureParameters, int priority) {
-        mImporter.loadTexture(resource, callback, null, priority, GVRCompressedTexture.BALANCED);
-    }
-
-    /**
-     * A simplified, low-level method that loads a texture asynchronously,
-     * without making you specify
-     * {@link #loadBitmapTexture(GVRAndroidResource.BitmapTextureCallback, GVRAndroidResource)
-     * loadBitmapTexture()} or
-     * {@link #loadCompressedTexture(GVRAndroidResource.CompressedTextureCallback, GVRAndroidResource)
-     * loadCompressedTexture()}.
-     * 
-     * This method can detect whether the resource file holds a compressed
-     * texture (GVRF currently supports ASTC, ETC2, and KTX formats:
-     * applications can add new formats by implementing
-     * {@link GVRCompressedTextureLoader}): if the file is not a compressed
-     * texture, it is loaded as a normal, bitmapped texture. This format
-     * detection adds very little to the cost of loading even a compressed
-     * texture, and it makes your life a lot easier: you can replace, say,
-     * {@code res/raw/resource.png} with {@code res/raw/resource.etc2} without
-     * having to change any code.
-     * 
-     * <p>
-     * We will continue to support the {@code loadBitmapTexture()} and
-     * {@code loadCompressedTexture()} APIs for at least a little while: We
-     * haven't yet decided whether to deprecate them or not.
-     * 
-     * @param callback
-     *            Before loading, GVRF may call
-     *            {@link GVRAndroidResource.TextureCallback#stillWanted(GVRAndroidResource)
-     *            stillWanted()} several times (on a background thread) to give
-     *            you a chance to abort a 'stale' load.
-     * 
-     *            Successful loads will call
-     *            {@link GVRAndroidResource.Callback#loaded(GVRHybridObject, GVRAndroidResource)
-     *            loaded()} on the GL thread;
-     * 
-     *            any errors will call
-     *            {@link GVRAndroidResource.TextureCallback#failed(Throwable, GVRAndroidResource)
-     *            failed()}, with no promises about threading.
-     * 
-     *            <p>
-     *            This method uses a throttler to avoid overloading the system.
-     *            If the throttler has threads available, it will run this
-     *            request immediately. Otherwise, it will enqueue the request,
-     *            and call
-     *            {@link GVRAndroidResource.TextureCallback#stillWanted(GVRAndroidResource)
-     *            stillWanted()} at least once (on a background thread) to give
-     *            you a chance to abort a 'stale' load.
-     * 
-     *            <p>
-     *            Use {@link #loadFutureTexture(GVRAndroidResource)} to avoid
-     *            having to implement a callback.
-     * @param resource
-     *            Basically, a stream containing a texture file. The
-     *            {@link GVRAndroidResource} class has six constructors to
-     *            handle a wide variety of Android resource types. Taking a
-     *            {@code GVRAndroidResource} here eliminates six overloads.
-     * @param textureParameters
-     *            The texture parameter object which has all the values that
-     *            were provided by the user for texture enhancement. The
-     *            {@link GVRTextureParameters} class has methods to set all the
-     *            texture filters and wrap states.
-     * @param priority
-     *            This request's priority. Please see the notes on asynchronous
-     *            priorities in the <a href="package-summary.html#async">package
-     *            description</a>. Also, please note priorities only apply to
-     *            uncompressed textures (standard Android bitmap files, which
-     *            can take hundreds of milliseconds to load): compressed
-     *            textures load so quickly that they are not run through the
-     *            request scheduler.
-     * @param quality
-     *            The compressed texture {@link GVRCompressedTexture#mQuality
-     *            quality} parameter: should be one of
-     *            {@link GVRCompressedTexture#SPEED},
-     *            {@link GVRCompressedTexture#BALANCED}, or
-     *            {@link GVRCompressedTexture#QUALITY}, but other values are
-     *            'clamped' to one of the recognized values. Please note that
-     *            this (currently) only applies to compressed textures; normal
-     *            {@linkplain GVRBitmapTexture bitmapped textures} don't take a
-     *            quality parameter.
-     * 
-     * @since 2.0.2
-     * 
-     * @throws IllegalArgumentException
-     *             If you 'abuse' request consolidation by passing the same
-     *             {@link GVRAndroidResource} descriptor to multiple load calls.
-     *             <p>
-     *             It's fairly common for multiple scene objects to use the same
-     *             texture or the same mesh. Thus, if you try to load, say,
-     *             {@code R.raw.whatever} while you already have a pending
-     *             request for {@code R.raw.whatever}, it will only be loaded
-     *             once; the same resource will be used to satisfy both (all)
-     *             requests. This "consolidation" uses
-     *             {@link GVRAndroidResource#equals(Object)}, <em>not</em>
-     *             {@code ==} (aka "reference equality"): The problem with using
-     *             the same resource descriptor is that if requests can't be
-     *             consolidated (because the later one(s) came in after the
-     *             earlier one(s) had already completed) the resource will be
-     *             reloaded ... but the original descriptor will have been
-     *             closed.
-     */
-    public void loadTexture(TextureCallback callback,
-            GVRAndroidResource resource,
-            GVRTextureParameters textureParameters, int priority, int quality) {
-        mImporter.loadTexture(resource, callback, textureParameters, priority, quality);
-    }
-
-    /**
-     * Simple, high-level method to load a texture asynchronously, for use with
-     * {@link GVRShaders#setMainTexture(Future)} and
-     * {@link GVRShaders#setTexture(String, Future)}.
-     * 
-     * This method uses a default priority and a default render quality: use
-     * {@link #loadFutureTexture(GVRAndroidResource, int)} to specify a priority
-     * or {@link #loadFutureTexture(GVRAndroidResource, int, int)} to specify a
-     * priority and render quality.
-     * 
-     * <p>
-     * This method is significantly easier to use than
-     * {@link #loadTexture(GVRAndroidResource.TextureCallback, GVRAndroidResource)}
-     * : you don't have to implement a callback; you don't have to pay attention
-     * to the low-level details of
-     * {@linkplain GVRSceneObject#attachRenderData(GVRRenderData) attaching} a
-     * {@link GVRRenderData} to your scene object. What's more, you don't even
-     * lose any functionality: {@link Future#cancel(boolean)} lets you cancel a
-     * 'stale' request, just like
-     * {@link GVRAndroidResource.CancelableCallback#stillWanted(GVRAndroidResource)
-     * stillWanted()} does. The flip side, of course, is that it <em>is</em> a
-     * bit more expensive: methods like
-     * {@link GVRMaterial#setMainTexture(Future)} use an extra thread from the
-     * thread pool to wait for the blocking {@link Future#get()} call. For
-     * modest numbers of loads, this overhead is acceptable - but thread
-     * creation is not free, and if your {@link GVRMain#onInit(GVRContext)
-     * onInit()} method fires of dozens of future loads, you may well see an
-     * impact.
-     * 
-     * @param resource
-     *            Basically, a stream containing a texture file. The
-     *            {@link GVRAndroidResource} class has six constructors to
-     *            handle a wide variety of Android resource types. Taking a
-     *            {@code GVRAndroidResource} here eliminates six overloads.
-     * @return A {@link Future} that you can pass to methods like
-     *         {@link GVRShaders#setMainTexture(Future)}
-     * 
-     * @since 1.6.7
-     * 
-     * @throws IllegalArgumentException
-     *             If you 'abuse' request consolidation by passing the same
-     *             {@link GVRAndroidResource} descriptor to multiple load calls.
-     *             <p>
-     *             It's fairly common for multiple scene objects to use the same
-     *             texture or the same mesh. Thus, if you try to load, say,
-     *             {@code R.raw.whatever} while you already have a pending
-     *             request for {@code R.raw.whatever}, it will only be loaded
-     *             once; the same resource will be used to satisfy both (all)
-     *             requests. This "consolidation" uses
-     *             {@link GVRAndroidResource#equals(Object)}, <em>not</em>
-     *             {@code ==} (aka "reference equality"): The problem with using
-     *             the same resource descriptor is that if requests can't be
-     *             consolidated (because the later one(s) came in after the
-     *             earlier one(s) had already completed) the resource will be
-     *             reloaded ... but the original descriptor will have been
-     *             closed.
-     */
-    public Future<GVRTexture> loadFutureTexture(GVRAndroidResource resource) {
-        return getAssetLoader().loadFutureTexture(resource);
-    }
-
-    /**
-     * Simple, high-level method to load a texture asynchronously, for use with
-     * {@link GVRShaders#setMainTexture(Future)} and
-     * {@link GVRShaders#setTexture(String, Future)}.
-     * 
-     * This method uses a default render quality:
-     * {@link #loadFutureTexture(GVRAndroidResource, int, int)} to specify
-     * render quality.
-     * 
-     * <p>
-     * This method is significantly easier to use than
-     * {@link #loadTexture(GVRAndroidResource.TextureCallback, GVRAndroidResource, int)}
-     * You don't have to implement a callback; you don't have to pay attention
-     * to the low-level details of
-     *{@linkplain GVRSceneObject#attachRenderData(GVRRenderData) attaching} a
-     * {@link GVRRenderData} to your scene object. What's more, you don't even
-     * lose any functionality: {@link Future#cancel(boolean)} lets you cancel a
-     * 'stale' request, just like
-     * {@link GVRAndroidResource.CancelableCallback#stillWanted(GVRAndroidResource)
-     * stillWanted()} does. The flip side, of course, is that it <em>is</em> a
-     * bit more expensive: methods like
-     * {@link GVRMaterial#setMainTexture(Future)} use an extra thread from the
-     * thread pool to wait for the blocking {@link Future#get()} call. For
-     * modest numbers of loads, this overhead is acceptable - but thread
-     * creation is not free, and if your {@link GVRMain#onInit(GVRContext)
-     * onInit()} method fires of dozens of future loads, you may well see an
-     * impact.
-     * 
-     * @param resource
-     *            Basically, a stream containing a texture file. The
-     *            {@link GVRAndroidResource} class has six constructors to
-     *            handle a wide variety of Android resource types. Taking a
-     *            {@code GVRAndroidResource} here eliminates six overloads.
-     * @param priority
-     *            This request's priority. Please see the notes on asynchronous
-     *            priorities in the <a href="package-summary.html#async">package
-     *            description</a>. Also, please note priorities only apply to
-     *            uncompressed textures (standard Android bitmap files, which
-     *            can take hundreds of milliseconds to load): compressed
-     *            textures load so quickly that they are not run through the
-     *            request scheduler.
-     * @return A {@link Future} that you can pass to methods like
-     *         {@link GVRShaders#setMainTexture(Future)}
-     * 
-     * @since 1.6.7
-     * 
-     * @throws IllegalArgumentException
-     *             If you 'abuse' request consolidation by passing the same
-     *             {@link GVRAndroidResource} descriptor to multiple load calls.
-     *             <p>
-     *             It's fairly common for multiple scene objects to use the same
-     *             texture or the same mesh. Thus, if you try to load, say,
-     *             {@code R.raw.whatever} while you already have a pending
-     *             request for {@code R.raw.whatever}, it will only be loaded
-     *             once; the same resource will be used to satisfy both (all)
-     *             requests. This "consolidation" uses
-     *             {@link GVRAndroidResource#equals(Object)}, <em>not</em>
-     *             {@code ==} (aka "reference equality"): The problem with using
-     *             the same resource descriptor is that if requests can't be
-     *             consolidated (because the later one(s) came in after the
-     *             earlier one(s) had already completed) the resource will be
-     *             reloaded ... but the original descriptor will have been
-     *             closed.
-     */
-    public Future<GVRTexture> loadFutureTexture(GVRAndroidResource resource,
-            int priority) {
-        return loadFutureTexture(resource, priority,
-                GVRCompressedTexture.BALANCED);
-    }
-
-    /**
-     * Simple, high-level method to load a texture asynchronously, for use with
-     * {@link GVRShaders#setMainTexture(Future)} and
-     * {@link GVRShaders#setTexture(String, Future)}.
-     * 
-     * 
-     * <p>
-     * This method is significantly easier to use than
-     * {@link #loadTexture(GVRAndroidResource.TextureCallback, GVRAndroidResource, int, int)}
-     * You don't have to implement a callback; you don't have to pay attention
-     * to the low-level details of
-     *{@linkplain GVRSceneObject#attachRenderData(GVRRenderData) attaching} a
-     * {@link GVRRenderData} to your scene object. What's more, you don't even
-     * lose any functionality: {@link Future#cancel(boolean)} lets you cancel a
-     * 'stale' request, just like
-     * {@link GVRAndroidResource.CancelableCallback#stillWanted(GVRAndroidResource)
-     * stillWanted()} does. The flip side, of course, is that it <em>is</em> a
-     * bit more expensive: methods like
-     * {@link GVRMaterial#setMainTexture(Future)} use an extra thread from the
-     * thread pool to wait for the blocking {@link Future#get()} call. For
-     * modest numbers of loads, this overhead is acceptable - but thread
-     * creation is not free, and if your {@link GVRMain#onInit(GVRContext)
-     * onInit()} method fires of dozens of future loads, you may well see an
-     * impact.
-     * 
-     * @param resource
-     *            Basically, a stream containing a texture file. The
-     *            {@link GVRAndroidResource} class has six constructors to
-     *            handle a wide variety of Android resource types. Taking a
-     *            {@code GVRAndroidResource} here eliminates six overloads.
-     * @param priority
-     *            This request's priority. Please see the notes on asynchronous
-     *            priorities in the <a href="package-summary.html#async">package
-     *            description</a>. Also, please note priorities only apply to
-     *            uncompressed textures (standard Android bitmap files, which
-     *            can take hundreds of milliseconds to load): compressed
-     *            textures load so quickly that they are not run through the
-     *            request scheduler.
-     * @param quality
-     *            The compressed texture {@link GVRCompressedTexture#mQuality
-     *            quality} parameter: should be one of
-     *            {@link GVRCompressedTexture#SPEED},
-     *            {@link GVRCompressedTexture#BALANCED}, or
-     *            {@link GVRCompressedTexture#QUALITY}, but other values are
-     *            'clamped' to one of the recognized values. Please note that
-     *            this (currently) only applies to compressed textures; normal
-     *            {@linkplain GVRBitmapTexture bitmapped textures} don't take a
-     *            quality parameter.
-     * @return A {@link Future} that you can pass to methods like
-     *         {@link GVRShaders#setMainTexture(Future)}
-     * 
-     * @since 1.6.7
-     * 
-     * @throws IllegalArgumentException
-     *             If you 'abuse' request consolidation by passing the same
-     *             {@link GVRAndroidResource} descriptor to multiple load calls.
-     *             <p>
-     *             It's fairly common for multiple scene objects to use the same
-     *             texture or the same mesh. Thus, if you try to load, say,
-     *             {@code R.raw.whatever} while you already have a pending
-     *             request for {@code R.raw.whatever}, it will only be loaded
-     *             once; the same resource will be used to satisfy both (all)
-     *             requests. This "consolidation" uses
-     *             {@link GVRAndroidResource#equals(Object)}, <em>not</em>
-     *             {@code ==} (aka "reference equality"): The problem with using
-     *             the same resource descriptor is that if requests can't be
-     *             consolidated (because the later one(s) came in after the
-     *             earlier one(s) had already completed) the resource will be
-     *             reloaded ... but the original descriptor will have been
-     *             closed.
-     */
-    public Future<GVRTexture> loadFutureTexture(GVRAndroidResource resource,
-            int priority, int quality) {
-        return mImporter.loadFutureTexture(resource, priority, quality);
-    }
-
-    /**
-     * Simple, high-level method to load a cube map texture asynchronously, for
-     * use with {@link GVRShaders#setMainTexture(Future)} and
-     * {@link GVRShaders#setTexture(String, Future)}.
-     * 
-     * @param resource
-     *            A steam containing a zip file which contains six bitmaps. The
-     *            six bitmaps correspond to +x, -x, +y, -y, +z, and -z faces of
-     *            the cube map texture respectively. The default names of the
-     *            six images are "posx.png", "negx.png", "posy.png", "negx.png",
-     *            "posz.png", and "negz.png", which can be changed by calling
-     *            {@link GVRCubemapTexture#setFaceNames(String[])}.
-     * @return A {@link Future} that you can pass to methods like
-     *         {@link GVRShaders#setMainTexture(Future)}
-     * 
-     * @since 1.6.9
-     * 
-     * @throws IllegalArgumentException
-     *             If you 'abuse' request consolidation by passing the same
-     *             {@link GVRAndroidResource} descriptor to multiple load calls.
-     *             <p>
-     *             It's fairly common for multiple scene objects to use the same
-     *             texture or the same mesh. Thus, if you try to load, say,
-     *             {@code R.raw.whatever} while you already have a pending
-     *             request for {@code R.raw.whatever}, it will only be loaded
-     *             once; the same resource will be used to satisfy both (all)
-     *             requests. This "consolidation" uses
-     *             {@link GVRAndroidResource#equals(Object)}, <em>not</em>
-     *             {@code ==} (aka "reference equality"): The problem with using
-     *             the same resource descriptor is that if requests can't be
-     *             consolidated (because the later one(s) came in after the
-     *             earlier one(s) had already completed) the resource will be
-     *             reloaded ... but the original descriptor will have been
-     *             closed.
-     */
-    public Future<GVRTexture> loadFutureCubemapTexture(
-            GVRAndroidResource resource) {
-        return mImporter.loadFutureCubemapTexture(resource);
-    }
-
-    /**
-     * Simple, high-level method to load a compressed cube map texture asynchronously,
-     * for use with {@link GVRShaders#setMainTexture(Future)} and
-     * {@link GVRShaders#setTexture(String, Future)}.
-     *
-     * @param resource
-     *            A steam containing a zip file which contains six compressed textures.
-     *            The six textures correspond to +x, -x, +y, -y, +z, and -z faces of
-     *            the cube map texture respectively. The default names of the
-     *            six images are "posx.pkm", "negx.pkm", "posy.pkm", "negx.pkm",
-     *            "posz.pkm", and "negz.pkm", which can be changed by calling
-     *            {@link GVRCubemapTexture#setFaceNames(String[])}.
-     * @return A {@link Future} that you can pass to methods like
-     *         {@link GVRShaders#setMainTexture(Future)}
-     *
-     * @since 1.6.9
-     *
-     * @throws IllegalArgumentException
-     *             If you 'abuse' request consolidation by passing the same
-     *             {@link GVRAndroidResource} descriptor to multiple load calls.
-     *             <p>
-     *             It's fairly common for multiple scene objects to use the same
-     *             texture or the same mesh. Thus, if you try to load, say,
-     *             {@code R.raw.whatever} while you already have a pending
-     *             request for {@code R.raw.whatever}, it will only be loaded
-     *             once; the same resource will be used to satisfy both (all)
-     *             requests. This "consolidation" uses
-     *             {@link GVRAndroidResource#equals(Object)}, <em>not</em>
-     *             {@code ==} (aka "reference equality"): The problem with using
-     *             the same resource descriptor is that if requests can't be
-     *             consolidated (because the later one(s) came in after the
-     *             earlier one(s) had already completed) the resource will be
-     *             reloaded ... but the original descriptor will have been
-     *             closed.
-     */
-    public Future<GVRTexture> loadFutureCompressedCubemapTexture(
-            GVRAndroidResource resource) {
-        return mImporter.loadFutureCompressedCubemapTexture(resource);
-    }
-
-    /**
-     * Loads atlas information file placed in the assets folder.
-     *
-     * Atlas information file contains in UV space the information of offset and
-     * scale for each mesh mapped in some atlas texture.
-     * The content of the file is at json format like:
-     *
-     * [ {name: SUN, offset.x: 0.9, offset.y: 0.9, scale.x: 0.5, scale.y: 0.5},
-     * {name: EARTH, offset.x: 0.5, offset.y: 0.9, scale.x: 0.5, scale.y: 0.5} ]
-     *
-     * @param resource
-     *            A stream containing a text file on JSON format.
-     *
-     * @return List of atlas information load.
-     */
-    public List<GVRAtlasInformation> loadTextureAtlasInformation(GVRAndroidResource resource) throws IOException {
-
-        List<GVRAtlasInformation> atlasInformation
-                = GVRAsynchronousResourceLoader.loadAtlasInformation(resource.getStream());
-        resource.closeStream();
-
-        return atlasInformation;
     }
 
     /**
