+++ GVRf/Framework/framework/src/main/java/org/gearvrf/GVRContextProxy.java
@@ -47,116 +47,9 @@ public final class GVRContextProxy extends GVRContext {
         return mContext.get().getEventReceiver();
     }
 
-    public GVRMesh loadMesh(GVRAndroidResource androidResource) {
-        return mContext.get().loadMesh(androidResource);
-    }
-
-    public GVRMesh loadMesh(GVRAndroidResource androidResource, EnumSet<GVRImportSettings> settings) {
-        return mContext.get().loadMesh(androidResource, settings);
-    }
-
-    public void loadMesh(GVRAndroidResource.MeshCallback callback, GVRAndroidResource androidResource) throws IllegalArgumentException {
-        mContext.get().loadMesh(callback, androidResource);
-    }
-
-    public void loadMesh(GVRAndroidResource.MeshCallback callback, GVRAndroidResource resource, int priority) throws IllegalArgumentException {
-        mContext.get().loadMesh(callback, resource, priority);
-    }
-
-    public Future<GVRMesh> loadFutureMesh(GVRAndroidResource resource) {
-        return mContext.get().loadFutureMesh(resource);
-    }
-
-    public Future<GVRMesh> loadFutureMesh(GVRAndroidResource resource, int priority) {
-        return mContext.get().loadFutureMesh(resource, priority);
-    }
-
-    public GVRSceneObject getAssimpModel(String assetRelativeFilename) throws IOException {
-        return mContext.get().getAssimpModel(assetRelativeFilename);
-    }
-
-    public GVRSceneObject getAssimpModel(String assetRelativeFilename, EnumSet<GVRImportSettings> settings) throws IOException {
-        return mContext.get().getAssimpModel(assetRelativeFilename, settings);
-    }
-
-    public GVRModelSceneObject loadJassimpModelFromSD(String externalFile) throws IOException {
-        return mContext.get().loadJassimpModelFromSD(externalFile);
-    }
-
-    public GVRModelSceneObject loadJassimpModelFromSD(String externalFile, EnumSet<GVRImportSettings> settings) throws IOException {
-        return mContext.get().loadJassimpModelFromSD(externalFile, settings);
-    }
-
-    public GVRModelSceneObject loadJassimpModel(String assetFile) throws IOException {
-        return mContext.get().loadJassimpModel(assetFile);
-    }
-
-    public GVRModelSceneObject loadJassimpModel(String assetFile, EnumSet<GVRImportSettings> settings) throws IOException {
-        return mContext.get().loadJassimpModel(assetFile, settings);
-    }
-
-    public GVRModelSceneObject loadModelFromSD(String externalFile) throws IOException {
-        return mContext.get().loadModelFromSD(externalFile);
-    }
-
-    public GVRModelSceneObject loadModelFromSD(String externalFile, EnumSet<GVRImportSettings> settings) throws IOException {
-        return mContext.get().loadModelFromSD(externalFile, settings);
-    }
-
-    public GVRModelSceneObject loadModel(String assetFile) throws IOException {
-        return mContext.get().loadModel(assetFile);
-    }
-
-    public GVRModelSceneObject loadModel(String assetFile, EnumSet<GVRImportSettings> settings, GVRScene scene) throws IOException {
-        return mContext.get().loadModel(assetFile, settings, scene);
-    }
-
-    public GVRModelSceneObject loadModel(String assetFile, EnumSet<GVRImportSettings> settings) throws IOException {
-        return mContext.get().loadModel(assetFile, settings);
-    }
-
-    public GVRSceneObject loadModelFromURL(String urlString) throws IOException {
-        return mContext.get().loadModelFromURL(urlString);
-    }
-
-    public GVRSceneObject loadModelFromURL(String urlString, boolean cacheEnabled) throws IOException {
-        return mContext.get().loadModelFromURL(urlString, cacheEnabled);
-    }
-
-    public GVRSceneObject loadModelFromURL(String urlString, EnumSet<GVRImportSettings> settings) throws IOException {
-        return mContext.get().loadModelFromURL(urlString, settings);
-    }
 
     public GVRMesh createQuad(float width, float height) {
         return mContext.get().createQuad(width, height);
-    }
-
-    public Bitmap loadBitmap(String fileName) {
-        return mContext.get().loadBitmap(fileName);
-    }
-
-    public GVRBitmapTexture loadTexture(String fileName) {
-        return mContext.get().loadTexture(fileName);
-    }
-
-    public GVRBitmapTexture loadTexture(String fileName, GVRTextureParameters textureParameters) {
-        return mContext.get().loadTexture(fileName, textureParameters);
-    }
-
-    public GVRTexture loadTexture(GVRAndroidResource resource) {
-        return mContext.get().loadTexture(resource);
-    }
-
-    public GVRTexture loadTexture(GVRAndroidResource resource, GVRTextureParameters textureParameters) {
-        return mContext.get().loadTexture(resource, textureParameters);
-    }
-
-    public GVRCubemapTexture loadCubemapTexture(GVRAndroidResource[] resourceArray) {
-        return mContext.get().loadCubemapTexture(resourceArray);
-    }
-
-    public GVRCubemapTexture loadCubemapTexture(GVRAndroidResource[] resourceArray, GVRTextureParameters textureParameters) {
-        return mContext.get().loadCubemapTexture(resourceArray, textureParameters);
     }
 
     public void assertGLThread() {
@@ -165,70 +58,6 @@ public final class GVRContextProxy extends GVRContext {
 
     public boolean isCurrentThreadGLThread() {
         return mContext.get().isCurrentThreadGLThread();
-    }
-
-    public void loadBitmapTexture(GVRAndroidResource.BitmapTextureCallback callback, GVRAndroidResource resource) {
-        mContext.get().loadBitmapTexture(callback, resource);
-    }
-
-    public void loadBitmapTexture(GVRAndroidResource.TextureCallback callback, GVRAndroidResource resource, int priority) throws IllegalArgumentException {
-        mContext.get().loadBitmapTexture(callback, resource, priority);
-    }
-
-    public void loadCompressedTexture(GVRAndroidResource.CompressedTextureCallback callback, GVRAndroidResource resource) {
-        mContext.get().loadCompressedTexture(callback, resource);
-    }
-
-    public void loadCompressedTexture(GVRAndroidResource.CompressedTextureCallback callback, GVRAndroidResource resource, int quality) {
-        mContext.get().loadCompressedTexture(callback, resource, quality);
-    }
-
-    public void loadTexture(GVRAndroidResource.TextureCallback callback, GVRAndroidResource resource) {
-        mContext.get().loadTexture(callback, resource);
-    }
-
-    public void loadTexture(GVRAndroidResource.TextureCallback callback, GVRAndroidResource resource, int priority) {
-        mContext.get().loadTexture(callback, resource, priority);
-    }
-
-    public void loadTexture(GVRAndroidResource.TextureCallback callback, GVRAndroidResource resource, int priority, int quality) {
-        mContext.get().loadTexture(callback, resource, priority, quality);
-    }
-
-    public void loadTexture(GVRAndroidResource.TextureCallback callback, GVRAndroidResource resource, GVRTextureParameters textureParameters) {
-        mContext.get().loadTexture(callback, resource, textureParameters);
-    }
-
-    public void loadTexture(GVRAndroidResource.TextureCallback callback, GVRAndroidResource resource, GVRTextureParameters textureParameters, int priority) {
-        mContext.get().loadTexture(callback, resource, textureParameters, priority);
-    }
-
-    public void loadTexture(GVRAndroidResource.TextureCallback callback, GVRAndroidResource resource, GVRTextureParameters textureParameters, int priority, int quality) {
-        mContext.get().loadTexture(callback, resource, textureParameters, priority, quality);
-    }
-
-    public Future<GVRTexture> loadFutureTexture(GVRAndroidResource resource) {
-        return mContext.get().loadFutureTexture(resource);
-    }
-
-    public Future<GVRTexture> loadFutureTexture(GVRAndroidResource resource, int priority) {
-        return mContext.get().loadFutureTexture(resource, priority);
-    }
-
-    public Future<GVRTexture> loadFutureTexture(GVRAndroidResource resource, int priority, int quality) {
-        return mContext.get().loadFutureTexture(resource, priority, quality);
-    }
-
-    public Future<GVRTexture> loadFutureCubemapTexture(GVRAndroidResource resource) {
-        return mContext.get().loadFutureCubemapTexture(resource);
-    }
-
-    public Future<GVRTexture> loadFutureCompressedCubemapTexture(GVRAndroidResource resource) {
-        return mContext.get().loadFutureCompressedCubemapTexture(resource);
-    }
-
-    public List<GVRAtlasInformation> loadTextureAtlasInformation(GVRAndroidResource resource) throws IOException {
-        return mContext.get().loadTextureAtlasInformation(resource);
     }
 
     public GVRScene getMainScene() {
