+++ GVRf/Framework/framework/src/main/java/org/gearvrf/GVRSceneObject.java
@@ -224,7 +224,7 @@ public class GVRSceneObject extends GVRHybridObject implements PrettyPrint, IScr
      */
     public GVRSceneObject(GVRContext gvrContext, GVRAndroidResource mesh,
             GVRAndroidResource texture) {
-        this(gvrContext, gvrContext.loadFutureMesh(mesh), gvrContext
+        this(gvrContext, gvrContext.getAssetLoader().loadFutureMesh(mesh), gvrContext
                 .getAssetLoader().loadFutureTexture(texture));
     }
